### 2021-02-20 #3048 @whedon
[PRE REVIEW]: IDCeMPy: Python Package for Inflated Discrete Choice Models
**Submitting author:** @hknd23 (<a href="http://orcid.org/0000-0002-6234-7232">Nguyen Huynh</a>)
**Repository:** <a href="https://github.com/hknd23/idcempy" target ="_blank">https://github.com/hknd23/idcempy</a>
**Version:** 0.0.1
