### 2021-02-26 #3048 @Sergio-Bejar
Hi @kthyng, thanks for keeping us in the loop!  We have no problem waiting until our submission is assigned to an editor.   Thanks to you and other members of the editorial board for your hard work.
