### 2021-05-19 #3048 @galessiorob
Happy to edit this paper @danielskatz, thanks for the invite!

👋 Hi @Sergio-Bejar

I am in the process of assigning reviewers to this submission, if you have any suggestions please let me know and feel free to @ mention them here.

Basic checks to go over before we formally start the review:

 - [ ] There’s a general description of the software
           Needs expanding
 - [ ] The software is within the scope as research software
           Potentially to be expanded to denote research application and use cases
 - [ ] OSI-approved license: GNU General Public License (GPL)
