### 2021-05-19 bced077
Delete old license
