+++ setup.py
@@ -21,7 +21,7 @@ setup(
     install_requires=["scipy>=1.4.1", "numpy", "pandas"],
     classifiers=[
         "Development Status :: 3 - Alpha",
-        "License :: OSI Approved :: MIT License",
+        "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
         "Programming Language :: Python :: 3.7",
         "Intended Audience :: Science/Research",
     ],
