### 2021-05-19 613d48d
Update setup.cfg
