+++ setup.cfg
@@ -10,7 +10,7 @@ long_description_content_type = text/markdown
 url = https://github.com/hknd23/idcempy
 classifiers =
 	Development Status :: 3 - Alpha
-	License :: OSI Approved :: MIT License
+	License :: OSI Approved :: GNU General Public License v3 (GPLv3)
         Programming Language :: Python :: 3.7
         Intended Audience :: Science/Research
 [options]
