+++ paper.md
@@ -33,7 +33,7 @@ affiliations:
 `IDCeMPy` is a free, open-source and versatile Python package that enables
 researchers to easily fit three distinct sets of inflated discrete choice models that
 are often used in a variety of disciplines including Economics, Business, Engineering, Political
-Ccience, Psychology, and Public Health: (i)Zero-Inflated Ordered Probit (ZIOP),(ii)
+Science, Psychology, and Public Health: (i)Zero-Inflated Ordered Probit (ZIOP),(ii)
 Middle-Inflated Ordered Probit (MIOP), and (iii)Generalized Inflated Multinomial
 Logit (GIMNL) models. While the ZIOP model permits careful evaluation of zero-
 inflated ordered choice outcomes that results from two data generating
@@ -51,57 +51,103 @@ documented. For further details, the reader is referred to the official `IDCeMPy
 repository.
 
 # Statement of Need
-Scholars and Data Scientists often use discrete choice models to evaluate ordered outcomes using the ordered 
-probit model and unordered polytomous measures via the multinomial logit (MNL) estimator. These models, 
-however, cannot account for the fact that in many ordered and unordered polytomous choice situations, 
-an excessive share of observations—stemming from two distinct d.g.p’s—fall into a single category which 
-is thus “inflated.” For instance, ordered outcome measures of self-reported smoking behavior that range 
-from 0 for “no smoking” to 3 for “smoking 20 cigarettes or more daily” contain excessive observations 
-in the zero (no smoking) category that includes two types of non-smokers: individuals who never smoke 
-cigarettes, and those who smoked previously but temporarily stopped smoking because of their high price 
-[@harris2007zero; @greene2015inflated].  
+Scholars and Data Scientists often use discrete choice models to evaluate
+ordered outcomes using the ordered probit model and unordered polytomous
+measures via the multinomial logit (MNL) estimator. These models, however,
+cannot account for the fact that in many ordered and unordered polytomous
+choice situations, an excessive share of observations—stemming from two
+distinct d.g.p’s—fall into a single category which is thus “inflated.” For instance,
+ordered outcome measures of self-reported smoking behavior that range from 0
+for “no smoking” to 3 for “smoking 20 cigarettes or more daily” contain excessive
+observations in the zero (no smoking) category that includes two types of non-
+smokers: individuals who never smoke cigarettes, and those who smoked
+previously but temporarily stopped smoking because of their high price
+[@harris2007zero; @greene2015inflated].
+In ordered choice measures such as immigration attitudes, it is the middle-
+category of “indifference” that is inflated since it includes respondents who are
+truly indifferent about immigration and those who select “indifference” because of
+social desirability reasons [@bagozzi2012mixture; @brown2020modelling].
+Further, in unordered polytomous variables of vote choice, for example, the
+baseline category is frequently inflated as it includes non-voters who abstain from
+voting in an election owing to temporary factors and routine non-voters who are 
+disengaged from the political process [@campbell2008religion; @bagozzi2017distinguishing]. 
+Failing to account for such category inflation in
+discrete choice measures leads to model misspecification, biased estimates, and
+incorrect inferences.
 
-In ordered choice measures such as immigration attitudes, it is the middle-category of “indifference” 
-that is inflated since it includes respondents who are truly indifferent about immigration and those 
-who select “indifference” because of social desirability reasons 
-[@bagozzi2012mixture; @brown2020modelling]. Further, in unordered polytomous variables of vote choice, 
-for example, the baseline category is frequently inflated as it includes non-voters who abstain from 
[…]
