+++ paper.bib
@@ -177,3 +177,10 @@
   doi = {10.1093/poq/nfn032}
 }
 
+@book{greene2010modeling,
+  title={Modeling ordered choices: A primer},
+  author={Greene, William H and Hensher, David A},
+  year={2010},
+  publisher={Cambridge University Press},
+  doi = {10.1017/CBO9780511845062
+}
