+++ paper.bib
@@ -182,5 +182,5 @@
   author={Greene, William H and Hensher, David A},
   year={2010},
   publisher={Cambridge University Press},
-  doi = {10.1017/CBO9780511845062
+  doi = {10.1017/CBO9780511845062}
 }
