+++ paper.md
@@ -33,13 +33,13 @@ affiliations:
 `IDCeMPy` is a free, open-source and versatile Python package that enables
 researchers to easily fit three distinct sets of inflated discrete choice models that
 are often used in a variety of disciplines including Economics, Business, Engineering, Political
-Science, Psychology, and Public Health: (i)Zero-Inflated Ordered Probit (ZIOP),(ii)
-Middle-Inflated Ordered Probit (MIOP), and (iii)Generalized Inflated Multinomial
-Logit (GIMNL) models. While the ZIOP model permits careful evaluation of zero-
-inflated ordered choice outcomes that results from two data generating
-processes (hereafter, “d.g.p’s”), the MIOP model accounts for ordered choice
-outcomes in which the inflated middle-category emerges from distinct d.g.p’s.
-The GIMNL models account for the large proportion—and heterogeneous
+Science, Psychology, and Public Health: (i) Zero-Inflated Ordered Probit (ZIOP),(ii)
+Middle-Inflated Ordered Probit (MIOP), and (iii) Generalized Inflated Multinomial
+Logit (GIMNL) models. While the ZIOP model [@harris2007zero] permits careful evaluation 
+of zero-inflated ordered choice outcomes that results from two data generating
+processes (hereafter, “d.g.p’s”), the MIOP model [@bagozzi2012mixture]
+accounts for ordered choice outcomes in which the inflated middle-category emerges 
+from distinct d.g.p’s. The GIMNL models [@bagozzi2017distinguishing] account for the large proportion—and heterogeneous
 mixture—of observations in the baseline and other unordered outcome
 categories within MNL models that evaluate multiple unordered polytomous
 outcomes.
