+++ paper.md
@@ -35,7 +35,7 @@ researchers to easily fit three distinct sets of inflated discrete choice models
 are often used in a variety of disciplines including Economics, Business, Engineering, Political
 Science, Psychology, and Public Health: (i) Zero-Inflated Ordered Probit (ZIOP),(ii)
 Middle-Inflated Ordered Probit (MIOP), and (iii) Generalized Inflated Multinomial
-Logit (GIMNL) models. While the ZIOP model [@harris2007zero; @greene2010modeling] permits careful evaluation 
+Logit (GIMNL) models. While the ZIOP model [@harris2007zero] permits careful evaluation 
 of zero-inflated ordered choice outcomes that results from two data generating
 processes (hereafter, “d.g.p’s”), the MIOP model [@bagozzi2012mixture]
 accounts for ordered choice outcomes in which the inflated middle-category emerges 
