### 2021-05-20 #3048 @Sergio-Bejar
👋 Hi @galessiorob,

Thanks for agreeing to edit our submission, and for providing us with suggestions to improve the quality of our manuscript!
We have addressed all your requests.  Specifically, (i) we expanded the general description of the software in both the Statement of Need and Package Architecture sections (see pages 2 & 3); (ii) expanded the description of the package's application and use cases in both the Summary, Package Architecture, and Application sections; and (iii) added the GNU General Public License (GPL) statement in both the paper and GitHub repo.

I went ahead and created a PDF file of the updated manuscript.  I hope this is ok with you.

We'd like to suggest the following people (GitHub user names) as potential reviewers:
- alexhanna
- ohmaddi
- cmaimone
- chainsawriot

Thanks again for your time and effort editing our submission.  Please let us know if there's any other information we should provide to start the formal review.
