### 2021-06-14 #3322 @jungtaekkim
Hi @hknd23,

I read your article and looked into the repository.

I think .gitignore file should contain the rules that exclude Python-related files such as __pycache__ and *.pyc. Please refer to the .gitignore template provided by GitHub or any other examples. Regarding this, idecempy/__pycache__ should be removed.

Moreover, in the article, please mention an installation guide or link to the online page for the installation guide.

It is a minor issue, but at Line 11, (ZIOP),(ii) -> (ZIOP),<SPACE>(ii).

Best,
Jungtaek.
