+++ paper.md
@@ -84,7 +84,7 @@ Python package that offers researchers the possibility of estimating and
 assessing the performance of inflated ordered probit models with and without
 correlated errors, the Middle-Inflated Ordered Probit with and without correlated errors,
 and inflated MNL models. `IDCeMPy` entirely depends on [Python](https://www.python.org/) standard libraries such as 
-`NumPy`, `pandas`, and `SciPy`, allowing it to be easily extensible, flexible and efficient time-wise which facilitates efficient handling of large datasets.  It is compatible with [Python](https://www.python.org/) 3.7+.
+`NumPy`, `pandas`, and `SciPy`, allowing it to be easily extensible, flexible and efficient time-wise which facilitates efficient handling of large datasets.  It is compatible with [Python](https://www.python.org/) 3.7+. An easy to follow [instalation guide](https://idcempy.readthedocs.io/en/latest/zmiopc_tutorial.html#installation) is availabe in the package's documentation website hosted by [ReadTheDocs](https://readthedocs.org/).
 
 # Package Architecture
 `IDCeMPy` provides functions to fit the Zero-Inflated Ordered Probit (ZIOP) model
@@ -123,7 +123,7 @@ assessed time-period analyzed in the survey questionnaire. The Zero-Inflated Ord
 Probit model incorporated in the `iopmod` and `iopcmod`functions in `IDCeMPy`
 permits researchers to account for excessive zeros stemming from two
 populations in ordered measures such as perceived levels of depression, 
-which leads to more accurate estimates.T he MIOP(C) models also estimated 
+which leads to more accurate estimates. The MIOP(C) models also estimated 
 using the `iopmod` and `iopcmod` functions can address middle-category inflation in, for example,
 ordered measures like monetary policy [@brown2020modelling] and attitudes
 towards European Union (EU) membership [@bagozzi2012mixture].
