### 2021-06-14 #3322 @hknd23
Dear @jungtaekkim,

Thank you very much for your feedback on our paper, package, and GitHub repository.  We have now addressed all the points that you raised, and you can find below a detailed response to each of your comments/suggestions.

1.  I think .gitignore file should contain the rules that exclude Python-related files such as __pycache__ directory and *.pyc. Please refer to the .gitignore template provided by GitHub or any other examples. Regarding this, idecempy/pycache should be removed.
Response:  This is a great comment.  You indeed are right in pointing out that the .gitignore file should contain the rules that exclude Python-related files.  Following your suggestion, we carefully looked at the GitHub's .gitignore template and then removed "idcempy/pycache" from the directory in our repository.

2.  Moreover, in the article, please mention an installation guide or link to the online page for the installation guide.
Response:  Thanks for this suggestion.  We addressed it in two different ways.  First, we created a subsection in the ReadTheDocs documentation file that provides users with easy to follow instructions to install our package (https://idcempy.readthedocs.io/en/latest/zmiopc_tutorial.html#installation). Second, we followed your suggestion and added a sentence in the paper that provides users with links to the installation guide in both ReadTheDocs and GitHub.

3.  It is a minor issue, but at Line 11, (ZIOP),(ii) -> (ZIOP),(ii).
Response:  We apologize for making this error.  We have fixed the typo.
Please let us know if you have any additional comments/suggestions.  Thanks again for your time and effort reviewing our submission.

Regards,
Nguyen
