### 2021-06-15 hknd23/idcempy#13 @cmaimone
Paper: efficiency claim
The end of the statement of need section says "efficient time-wise" -- this reads like a claim about the performance of your package/code - either remove/rephrase this or provide performance evaluation information.
