### 2021-06-15 hknd23/idcempy#16 @cmaimone
community guidelines missing
From the JOSS review criteria:

Community guidelines

There should be clear guidelines for third-parties wishing to:

    Contribute to the software
    Report issues or problems with the software
    Seek support

This can be done via a contributing file in GitHub - see https://docs.github.com/en/communities/setting-up-your-project-for-healthy-contributions/setting-guidelines-for-repository-contributors
