### 2021-06-15 hknd23/idcempy#17 @cmaimone
Paper Citations: cite packages you depend on
Paper should provide citations for the packages you depend on - we are doing all this in part to make sure software we use gets cited :)
