### 2021-06-15 hknd23/idcempy#18 @cmaimone
Paper: State of the Field question
Are the two cited Stata commands the only implementations of these models in software?  Are there relevant R or Python packages for these models?  Either discuss/cite such packages or note that there are no implementations in R and Python to date.
