### 2021-06-15 hknd23/idcempy#19 @cmaimone
Testing: tests for correctness, coverage
The testing included currently tests that some of the main functions can be called and the code executes without error.  Tests should be expanded to:

1. Test that all of the functions execute without error.
2. Add tests to verify that the functions work correctly, not just without error -- for example, can you assert what the coefficients should be when you fit a particular model?  Or at least how many coefficients there should be?  There are multiple ways you may get at whether the output is correct.

Tests for the correctness of the results are needed so that you know if future changes to your code change the output in anyway.  As this could have a significant impact for the users (could change reported research results), this is important.

If there are published examples of what the model should produce for various inputs (for example, in the papers you cite), those would be good cases to use for testing -- verify that your code produces the same results as the published algorithms.
