### 2021-06-15 #3322 @cmaimone
Update:

I've added a number of issues and a PR for small formatting issues.

Notable issues, related to the checklist:

Performance: https://github.com/hknd23/idcempy/issues/13

Documentation: there are multiple issues opened for specific concerns.  Overall, the documentation could use work around:
1) API documentation is not very useful in its current format because the parameter descriptions are too vague
2) Organization of the documentation -- structure and titles of pages, making examples independent
3) Another round of proofreading throughout

Community Guidelines: https://github.com/hknd23/idcempy/issues/16

References: https://github.com/hknd23/idcempy/issues/17

State of the Field: https://github.com/hknd23/idcempy/issues/18

Testing: https://github.com/hknd23/idcempy/issues/19

Scholarly Effort: This package will be used and cited.  But it doesn't appear to have been yet.  It has a minimal history (code written over about a month period) and no significant history of use by people other than the authors.  The only significant use so far (that I can tell) is in creating the examples in the documentation.  I would prefer to see it go through a period of use and issue/bug reporting before being published, but I defer this to an editorial decision.
