### 2021-06-15 dc0538e
Delete gimnl_tutorial.rst
