+++ docs/zmiopc_tutorial.rst
@@ -58,12 +58,11 @@ The package can be installed in two different ways.
 Examples
 ========
 
-Zero-inflated Ordered Probit Models without Correlated Errors (ZiOP)
+Zero-inflated Ordered Probit without Correlated Errors (ZiOP)
 --------------------------------------------------------------------
 The `iopmod` function estimates regression objects for "zero-inflated" and "middle-inflated" ordered probit models without correlated errors.  Below you will find instructions to estimate a ZiOP model.  The estimation of MiOP models is also illustrated.
 
-
-**1. Import the required libraries, set up the package and import the dataset:**
+First, we import the required libraries, set up the package and import the dataset:
 
 .. testcode::
   # Import the necessary libraries and package
@@ -73,24 +72,23 @@ The `iopmod` function estimates regression objects for "zero-inflated" and "midd
   import urllib
   from zmiopc import zmiopc
 
-  # Import the "Youth Tobacco Consumption" dataset described above
+  # Import the "Youth Tobacco Consumption" dataset.
 
   url='https://github.com/hknd23/zmiopc/blob/main/data/tobacco_cons.csv'
   data=pd.read_stata(url)
 
-**2. Estimation of the ZiOP model.**
+Our data is now a `pandas` DataFrame, and we can proceed to estimate the ZiOP model as follows.
 
 .. testcode::
 
-  # Define a list of variable names (strings) X,Y,Z:
+  # First, you should define a list of variable names of X, Z, and Y.
+  # X = The covariates of the ordered probit stage.
+  # Z = The covariates of the inflation (split-population) stage.
+  # Y = The ordinal outcome variable.
+
   X = ['age', 'grade', 'gender_dum']
   Z = ['gender_dum']
   Y = ['cig_count']
-
-X is the list of variables in the Ordered Probit equation (second-stage).
-Z is the list of variables in the split-probit equation (first-stage).
-Y is the outcome variable.
-
 
 The package sets a default start value of .01 for all parameters.  Users can modify it by creating an array with their desired values, define such array as `pstart` and add it to as an argument in the model function.  
 
@@ -175,9 +173,79 @@ An example for the AIC:
  [0.87603805 0.06808193 0.01543795 0.02735256 0.01308951]
  [0.82681957 0.08778215 0.02153509 0.04095753 0.02290566]]
 
+Zero-inflated Ordered Probit With Correlated Errors (ZiOPC)
 
-**3. Estimation of the MiOP model**
+The package also includes the function `iopcmod` which fits "zero-inflated" ordered probit models (ZiOPC) and "middle-inflated" ordered probit models (MiOP) under the assumption that the two errors are correlated with each other (i.e. correlated errors). Both models include the estimate of'rho'. The models in this section use the same specification as the models estimated without correlated errors presented above.
 
+
+**1. Estimate the ZiOPC model**
+.. testcode::
+
+    ziopc_tob = zmiopc.iopcmod('ziopc', data, X, Y, Z, method='bfgs', weights=1, offsetx=0, offsetz=0)
+
+Similar to ZiOP, the results are stored in the attributes of :class:`zmiopc.IopCModel`.
+
+.. testoutput::
+
+         Current function value: 5060.051910
+         Iterations: 119
+         Function evaluations: 1562
+         Gradient evaluations: 142
+
+**2. Print the results**
+
+.. testcode::
+
+    print(ziopc_tob.coefs)
[…]
