+++ docs/zmiopc_tutorial.rst
@@ -77,7 +77,9 @@ We first import the required libraries, set up the package and import the datase
   # Import the "Youth Tobacco Consumption" dataset.
 
   url='https://github.com/hknd23/zmiopc/blob/main/data/tobacco_cons.csv'
-  data=pd.read_stata(url)
+
+  # Read the dataset
+  data=pd.read_csv(url)
 
 Our data is now a `pandas` DataFrame, and we can proceed to estimate the ZiOP model as follows.
 
@@ -179,7 +181,7 @@ The package also includes the function `iopcmod` which fits "zero-inflated" orde
 We first import the required libraries, set up the package and import the dataset:
 
 .. testcode::
-  # Import the necessary libraries and package
+  # Import the necessary libraries and IDCeMPy.
 
   import numpy as np
   import pandas as pd
@@ -189,6 +191,8 @@ We first import the required libraries, set up the package and import the datase
   # Import the "Youth Tobacco Consumption" dataset.
 
   url='https://github.com/hknd23/zmiopc/blob/main/data/tobacco_cons.csv'
+
+  # Read the imported dataset.
   data=pd.read_stata(url)
 
 .. testcode::
@@ -274,34 +278,48 @@ The predicted probabilities from the `ziopc_tob` model can ve obtained as follow
 
 Middle-inflated Ordered Probit without Correlated Errors (MiOP)
 ---------------------------------------------------------------
-We begin by importing the Elgun and Tilam (`2007 <https://journals.sagepub.com/doi/10.1177/1065912907305684>`_) data on European Integration described above.  Recall that our outcome variable is "inflated" in the middle category.
+If your ordered outcome variable is inflated in the middle category, you should estimate a  Middle-inflated Ordered Probit (MiOP) model.
+
+The following example uses data from Elgun and Tilam (`2007 <https://journals.sagepub.com/doi/10.1177/1065912907305684>`_).
+
+We begin by loading the required libraries and `IDCeMPy`
 
 .. testcode::
+  # Import the necessary libraries and IDCeMPy.
 
+  import numpy as np
+  import pandas as pd
+  import urllib
+  from zmiopc import zmiopc
+
+Next, we load the dataset.
+
+.. testcode::
+    # Import and read the dataset
     url = 'https://github.com/hknd23/zmiopc/blob/main/data/'
     data2 = pd_read.stata(url)
 
 We then define the lists with the names of the variables used in the model
 
 .. testcode::
+  # X = The covariates of the ordered probit stage.
+  # Z = The covariates of the inflation (split-population) stage.
+  # Y = The ordinal outcome variable.
 
   X = ['Xenophobia', 'discuss_politics']
   Z = ['discuss_politics', EU_Know_ob]
   Y = ['EU_support_ET']
 
-X is the list of variables in the Ordered Probit equation (second-stage).
-Z is the list of variables in the split-probit equation (first-stage).
-Y is the outcome variable.
-
+Your data is now ready, and you can begin the estimation process.
 
 :func:`zmiopc.iopmod` estimates the MiOP model and returns :class:`zmiopc.IopModel`.
 
 .. testcode::
 
-
   # Model estimation:
   miop_EU = zmiopc.iopmod('miop', data, X, Y, Z, method='bfgs', weights= 1,offsetx= 0, offsetz=0)
 
+The following message will appear when the model finishes converging.
 .. testoutput::
 
          Warning: Desired error not necessarily achieved due to precision loss.
@@ -310,9 +328,10 @@ Y is the outcome variable.
          Function evaluations: 488
          Gradient evaluations: 61  # See estimates:
 
+Print the results.
 .. testcode::
 
-         print(miop_EU.coefs)
[…]
