+++ docs/zmiopc_tutorial.rst
@@ -33,10 +33,6 @@ An excessive (“inflated”) share of observations—stemming from two distinct
 
 Each inflated discrete choice model in this package addresses category inflation in one’s discrete outcome—unordered or unordered polytomous—of interest by jointly estimating a binary split-stage equation and an ordered or multinomial discrete choice outcome equation.
 
-How does it work?
-=================
-
-
 Installation
 =============
 The package can be installed in two different ways.
@@ -72,7 +68,7 @@ We first import the required libraries, set up the package and import the datase
   import numpy as np
   import pandas as pd
   import urllib
-  from zmiopc import zmiopc
+  from idcempy import zmiopc
 
   # Import the "Youth Tobacco Consumption" dataset.
 
@@ -161,8 +157,8 @@ You can also extract predicted probabilities from the model:
 
 .. testcode::
 
-  fitttedziop = ziopc.iopfit(ziop_tob)
-  print(fitttedziopc.responsefull)
+  fittedziop = ziopc.iopfit(ziop_tob)
+  print(fittedziopc.responsefull)
 
 .. testoutput::
 
@@ -173,6 +169,23 @@ You can also extract predicted probabilities from the model:
  [0.73347708 0.1291157  0.03295816 0.06500889 0.03944016]
  [0.87603805 0.06808193 0.01543795 0.02735256 0.01308951]
  [0.82681957 0.08778215 0.02153509 0.04095753 0.02290566]]
+
+:func:`zmiopc.split_effects` simulates data from ZiOP results and computes changes in predicted probabilities when the value of a variable changes.
+This allows you to illustrate how the changes in the split-probit covariates affect the probabilities of being in one population versus another. The example below illustrates the marginal effects of the variable 'gender_dum' on the outcome variable in the ZiOPC model estimated in ths documentation.
+
+.. testcode::
+
+    ziopcgender = idcempy.split_effects(ziop_tob, 1, nsims = 10000)
+
+The returned dataframe contains predicted probabilities when 'gender_dum' equals 0, and when 'gender_dum' equals 1.
+The box plots below illustrate the change in predicted probabilities using the values from the 'ziopparl' dataframe.
+
+:func:: `zmiopc.ordered_effects` calculates the change in predicted probabilities of the outcome variable when the value of a covarariate changes. The box plots below display the change in predicted probabilities of the outcome variable in the ZiOP model.
+
+.. testcode::
+
+    gender = zmiopc.ordered_effects(ziop_tob, 2, nsims = 10000)
+    gender.plot.box(grid='False')
 
 Zero-inflated Ordered Probit (ZiOPC) with Correlated Errors
 -----------------------------------------------------------
@@ -186,7 +199,7 @@ We first import the required libraries, set up the package and import the datase
   import numpy as np
   import pandas as pd
   import urllib
-  from zmiopc import zmiopc
+  from idcempy import zmiopc
 
   # Import the "Youth Tobacco Consumption" dataset.
 
@@ -263,8 +276,8 @@ The predicted probabilities from the `ziopc_tob` model can ve obtained as follow
 
 .. testcode::
 
-  fitttedziopc = zmiopc.iopcfit(ziopc_tob)
-  print(fitttedziopc.responsefull)
+  fittedziopc = zmiopc.iopcfit(ziopc_tob)
+  print(fittedziopc.responsefull)
 
 .. testoutput::
 
@@ -276,13 +289,34 @@ The predicted probabilities from the `ziopc_tob` model can ve obtained as follow
  [0.87523652 0.06888286 0.01564958 0.0275354  0.01269564]
  [0.82678185 0.0875059  0.02171135 0.04135142 0.02264948]]
 
-Middle-inflated Ordered Probit without Correlated Errors (MiOP)
[…]
