+++ docs/zmiopc_tutorial.rst
@@ -382,13 +382,14 @@ Print the results.
 
 In addition to coefficient estimates, the table also presents the standard errors, and confidence intervals.
 
-The model object also stores three different diagnostic tests: (1) Log-likelihood, (2) Akaike Information Criteria (AIC), and Variance-Covariance Matrix (VCM).  You can obtain them via the following commands:
+The model object also stores three different diagnostic tests: (1) Log-likelihood, (2) Akaike Information Criteria (AIC), and Variance-Covariance Matrix (VCM).
 
 .. testcode::
-  # Print estiimates of LL, AIC and VCOV
-  print(miop_EU.llik)
-  print(miop_EU.AIC)
-  print(miop_EU.vcov)
+
+   # Print estimates of LL, AIC and VCOV
+   print(miop_EU.llik)
+   print(miop_EU.AIC)
+   print(miop_EU.vcov)
 
 For example, the AIC in this case is:
 
@@ -404,8 +405,8 @@ To estimate the predicted probabilities:
 
 .. testcode::
 
-  fittedmiop = zmiopc.iopcfit(miop_EU)
-  print(fittedziopc.responsefull)
+   fittedmiop = zmiopc.iopcfit(miop_EU)
+   print(fittedziopc.responsefull)
 
 The package also allows you to simulates data from MiOP model results and compute changes in predicted probabilities when the value of a variable changes.
 This allows you to illustrate how the changes in the split-probit covariates affect the probablilities of being in one population versus another.
@@ -445,6 +446,7 @@ We begin by loading the required libraries and IDCeMPy
 Next, we load the dataset.
 
 .. testcode::
+
     # Import and read the dataset
     url = 'https://github.com/hknd23/zmiopc/blob/main/data/'
     data2 = pd_read.stata(url)
@@ -452,13 +454,13 @@ Next, we load the dataset.
 We then define the lists with the names of the variables used in the model
 
 .. testcode::
-  # X = The covariates of the ordered probit stage.
-  # Z = The covariates of the inflation (split-population) stage.
-  # Y = The ordinal outcome variable.
 
-  X = ['Xenophobia', 'discuss_politics']
-  Z = ['discuss_politics', EU_Know_ob]
-  Y = ['EU_support_ET']
+   # X = The covariates of the ordered probit stage.
+   # Z = The covariates of the inflation (split-population) stage.
+   # Y = The ordinal outcome variable.
+   X = ['Xenophobia', 'discuss_politics']
+   Z = ['discuss_politics', EU_Know_ob]
+   Y = ['EU_support_ET']
 
 The model can be estimated as follows.
 
@@ -466,10 +468,8 @@ The model can be estimated as follows.
 
 .. testcode::
 
-  # Model estimation:
-  miopc_EU = zmiopc.iopcmod('miopc', pstartziop, data, X, Y, Z, method='bfgs', weights= 1,offsetx= 0, offsetz=0)
-
-.. testcode::
+   # Model estimation:
+   miopc_EU = zmiopc.iopcmod('miopc', pstartziop, data, X, Y, Z, method='bfgs', weights= 1,offsetx= 0, offsetz=0)
 
 Now print(miopc_EU.coefs).
 
@@ -491,16 +491,16 @@ The model object also stores three different diagnostic tests: (1) Log-likelihoo
 
 .. testcode::
 
-  print(miopc_EU.llik)
-  print(miopc_EU.AIC)
-  print(miopc_EU.vcov)
+   print(miopc_EU.llik)
+   print(miopc_EU.AIC)
+   print(miopc_EU.vcov)
 
 To estimate the predicted probabilities:
 
 .. testcode::
 
-  fittedmiopc = zmiopc.iopcfit(miopc_EU)
[…]
