+++ docs/index.rst
@@ -3,9 +3,6 @@
 Package Documentation: IDCeMPy
 ==============================
 
-The zmiopc package implement mixture models for data in which the *ordinal* response variable is characterized by an *excessive* number of observations in one of the ordinal outcomes.
-The package supports estimation of the Zero-inflated Ordered Probit and Middle-inflated Ordered Probit Models.
-
 Guide
 ^^^^^
 
