### 2021-06-16 180314f
some edits to docstring, new file
