+++ docs/idcempy_tutorial.rst
@@ -39,6 +39,7 @@ The package can be installed in two different ways.
 1. From `PyPi <https://pypi.org/>`__:
 
 .. testcode::
+
   # Import the package
 
   pip install idcempy
@@ -46,6 +47,7 @@ The package can be installed in two different ways.
 2. From its `GitHub Repository <https://github.com/hknd23/idcempy/>`__:
 
 .. testcode::
+
   # Import the package
 
   git clone https://github.com/hknd23/idcempy.git
@@ -62,6 +64,7 @@ The `iopcmod` function estimates regression objects for "zero-inflated" and "mid
 We first import the required libraries, set up the package and import the dataset:
 
 .. testcode::
+
   # Import the necessary libraries and package
 
   import numpy as np
@@ -194,6 +197,7 @@ The package also includes the function `iopcmod` which fits "zero-inflated" orde
 We first import the required libraries, set up the package and import the dataset:
 
 .. testcode::
+
   # Import the necessary libraries and IDCeMPy.
 
   import numpy as np
@@ -300,6 +304,7 @@ You can calculate the change in predicted probabilities of the outcome variable 
 In addition, a you can obtain a box plot that displays the change in predicted probabilities of the outcome variable in the ZiOPC model.
 
 .. testcode::
+
    # Calculate change in predicted probabilities
    gender = zmiopc.ordered_effects(ziopc_tob, 1, nsims = 10000)
 
@@ -315,6 +320,7 @@ The following example uses data from Elgun and Tilam (`2007 <https://journals.sa
 We begin by loading the required libraries and IDCeMPy
 
 .. testcode::
+
   # Import the necessary libraries and IDCeMPy.
 
   import numpy as np
@@ -323,14 +329,18 @@ We begin by loading the required libraries and IDCeMPy
   from idcempy import zmiopc
 
 Next, we load the dataset.
+
 .. testcode::
+
     # Import and read the dataset
     url = 'https://github.com/hknd23/zmiopc/blob/main/data/'
     # Define a `Pandas` DataFrame
     data = pd_read.stata(url)
 
 We then define the lists with the names of the variables used in the model
+
 .. testcode::
+
   # First, you should define a list of variable names of X, Z, and Y.
   # X = Column names of covariates (from `DataFrame`) used in ordered probit stage.
   # Z = Column names of covariates (from `DataFrame`) used in split-population stage.
@@ -360,6 +370,7 @@ The following message will appear when the model finishes converging.
          Gradient evaluations: 61  # See estimates:
 
 Print the results.
+
 .. testcode::
 
    print(miop_EU.coefs)
@@ -431,6 +442,7 @@ You can estimate a Middle-inflated Ordered Probit (MiOPC) with correlated errors
 We begin by loading the required libraries and IDCeMPy
 
 .. testcode::
+
   # Import the necessary libraries and IDCeMPy.
 
   import numpy as np
@@ -450,6 +462,7 @@ Next, we load the dataset.
 We then define the lists with the names of the variables used in the model
 
 .. testcode::
+
    # First, you should define a list of variable names of X, Z, and Y.
    # X = Column names of covariates (from `DataFrame`) used in ordered probit stage.
    # Z = Column names of covariates (from `DataFrame`) used in split-population stage.
@@ -530,6 +543,7 @@ The OP model does not account for the "zero inflation", so it does not have a sp
 We first import the required libraries, set up the package and import the dataset:
 
 .. testcode::
+
    # Import the necessary libraries and package
 
    import numpy as np
@@ -638,6 +652,7 @@ Failing to account for such inflation could lead to inaccurate inferences.
[…]
