### 2021-06-16 74dfcff
some edits to docstrings
