+++ idcempy/gimnl.py
@@ -250,7 +250,7 @@ def gimnlresults(model, data, x, y, z, modeltype, reference, inflatecat):
     :py:class:`GimnlModel`.
 
     :param model: object model estimated.
-    :param data: dataset.
+    :param data: dataset with missing values omitted.
     :param x: Multinomial Logit stage covariates.
     :param y: Dependent Variable (DV).
     :param z: Spplit-stage covariates.
@@ -331,10 +331,8 @@ def mnlresults(model, data, x, y, modeltype, reference):
     :param data: dataset.
     :param x: Multinomial Logit stage covariates.
     :param y: Dependent Variable (DV).
-    :param z: Spplit-stage covariates.
     :param modeltype: type of inflated MNL model.
     :param reference: order of categories.
-    :param inflatecat: inflated category.
     """
     varlist = np.unique(y + x)
     dataset = data[varlist]
@@ -399,10 +397,15 @@ def gimnlmod(data, x, y, z, reference, inflatecat, method="BFGS", pstart=None):
     :param y: Dependent Variable. Variable needs to be in factor form,
             with a number from 0-2 representing each category.
     :param z: Inflation stage covariates.
-    :param reference: order of categories.
-    :param inflatecat: inflated category. One of "baseline", "second",
-         or "third."
-    :param method: Optimization method.  Default is 'BFGS'
+    :param reference:  List specifying the order of categories (e.g [0, 1,
+    2], [2, 1, 0]. etc...). The parameter inflatecat then specifies which
+    category in the list inflated.
+    :param inflatecat: inflated category. One of "baseline" for the first,
+    "second" for the second,
+         or "third" for the third in reference list to specify the inflated
+         category.
+    :param method: Optimization method.  Default is 'BFGS'. For other
+    available methods, see scipy.optimize.minimize documentation.
     :param pstart: Starting parameters.
     """
     varlist = np.unique(y + z + x)
@@ -466,8 +469,10 @@ def mnlmod(data, x, y, reference, method="BFGS", pstart=None):
     :param x: MNL stage covariates.
     :param y: Dependent Variable. Variable needs to be in factor form,
          with a number from 0-2 representing each category.
-    :param reference: order of categories.
-    :param method: Optimization method.  Default is 'BFGS'
+    :param reference: order of categories. List specifying the order of
+    categories (e.g [0, 1, 2], [2, 1, 0]. etc...)
+    :param method: Optimization method.  Default is 'BFGS'. For other
+    available methods, see scipy.optimize.minimize documentation.
     :param pstart: Starting parameters.
     """
     varlist = np.unique(y + x)
