+++ idcempy/zmiopc.py
@@ -572,7 +572,8 @@ def opmod(data, x, y, pstart=None, method="BFGS", weights=1, offsetx=0):
     :type x: list of str
     :param y: Dependent Variable (DV).
     :type y: list of str
-    :param method: method for optimization, default 'BFGS'.
+    :param method: method for optimization, default 'BFGS'. For other
+    available methods, see scipy.optimize.minimize documentation.
     :param weights: weights.
     :param offsetx: offset for X.
     :return: OpModel
@@ -760,7 +761,8 @@ def iopmod(
     :param z: Inflation stage variables.
     :type z: list of str.
     :param modeltype: must be one of "ziop" or 'miop'.
-    :param method: method for optimization, default 'BFGS'.
+    :param method: method for optimization, default 'BFGS'. For other
+    available methods, see scipy.optimize.minimize documentation.
     :param weights: weights.
     :param offsetx: offset for X.
     :param offsetz: offset for Z.
@@ -827,7 +829,8 @@ def iopcmod(
     :param z: Inflation stage variables.
     :type z: list of str
     :param modeltype: Type of model to be estimated ("ziopc" or 'miopc').
-    :param method: method for optimization, default 'BFGS'
+    :param method: method for optimization, default 'BFGS'.  For other
+    available methods, see scipy.optimize.minimize documentation.
     :param weights: weights.
     :param offsetx: offset for X.
     :param offsetz: offset for Z.
