### 2021-06-16 e44dec3
update docstrings for API
