### 2021-06-16 f6b1d92
edit docstrings
