+++ idcempy/gimnl.py
@@ -40,8 +40,8 @@ class GimnlModel:
         :param aic: Model Akaike information .
         :param vcov: Variance-Covariance matrix.
             (optimized as inverted Hessian matrix)
-        :param data: Full dataset used for estimation, with missing values
-        dropped.
+        :param data: Model data used for estimation, subsetted to selected
+            variables and missing values listwise deleted.
         :param zs: Inflation stage estimates (Gammas).
         :param xs: Multinomial Logit probit estimates (Betas).
         :param ycatu: Number of categories in the Dependent Variable (DV).
@@ -102,8 +102,8 @@ class MnlModel:
         :param aic: Model Akaike information .
         :param vcov: Variance-Covariance matrix.
             (optimized as inverted Hessian matrix)
-        :param data: Full dataset used in estimation, with missing values
-        dropped.
+        :param data: Model data used for estimation, subsetted to selected
+            variables and missing values listwise deleted.
         :param xs: Multinomial Logit estimates (Betas).
         :param ycatu: Number of categories in the Dependent Variable (DV).
         :param x_: X Data.
@@ -130,11 +130,17 @@ def mnl3(pstart, x2, x3, y, reference):
     """
     Likelihood function for the baseline inflated three-category MNL model.
 
-    :param pstart: starting parameters.
-    :param x2: X (Multinomial Logit) covariates.
-    :param x3: X (Multinomial Logit) covariates (should be identical to x2).
-    :param y: Dependent Variable (DV).
-    :param reference: order of categories.
+    :param pstart: A list of starting parameters.
+    :type pstart: list
+    :param x2: Multinomial Logit covariates. Data subsetted to selected
+        variables.
+    :type x2: pandas dataframe
+    :param x3: Multinomial Logit covariates (should be identical to x2.
+    :type x3: pandas dataframe
+    :param y: Dependent Variable (DV). Data subsetted to selected
+        variable.
+    :type y: pandas dataframe
+    :param reference: list of order of categories.
     """
     b2 = pstart[0: len(x2.columns)]
     b3 = pstart[len(x2.columns): (len(pstart))]
@@ -156,12 +162,21 @@ def bimnl3(pstart, x2, x3, y, z, reference):
     """
     Likelihood function for the baseline inflated three-category MNL model.
 
-    :param pstart: starting parameters.
-    :param x2: X (Multinomial Logit) covariates.
-    :param x3: X (Multinomial Logit) covariates (should be identical to x2.
-    :param y: Dependent Variable (DV).
-    :param z: Logit Split stage covariates.
-    :param reference: order of categories (first category/baseline inflated).
+    :param pstart: A list of starting parameters.
+    :type pstart: list
+    :param x2: Multinomial Logit covariates. Data subsetted to selected
+        variables.
+    :type x2: pandas dataframe
+    :param x3: Multinomial Logit covariates (should be identical to x2.
+    :type x3: pandas dataframe
+    :param y: Dependent Variable (DV). Data subsetted to selected
+        variable.
+    :type y: pandas dataframe
+    :param z: Logit Split stage covariates. Data subsetted to selected
+        variables.
+    :type z: pandas dataframe
+    :param reference: list of order of categories (first element will be the
+        inflated category).
     """
     b2 = pstart[len(z.columns): (len(z.columns) + len(x2.columns))]
     b3 = pstart[(len(z.columns) + len(x2.columns)): (len(pstart))]
@@ -184,14 +199,23 @@ def bimnl3(pstart, x2, x3, y, z, reference):
 
 def simnl3(pstart, x2, x3, y, z, reference):
     """
-    Likelihood function for the second category inflated MNL model.
+    Likelihood function for the baseline inflated three-category MNL model.
 
-    :param pstart: starting parameters.
-    :param x2: X covariates.
[…]
