+++ idcempy/zmiopc.py
@@ -189,14 +189,14 @@ class FittedVals:
 def op(pstart, x, y, data, weights, offsetx):
     """Calculate likelihood function for Ordered Probit Model.
 
-    :param pstart: starting parameters.
-    :type pstart: numpy.ndarray or list.
+    :param pstart: A list of starting parameters.
+    :type pstart: list
     :param x: Ordered stage variables.
-    :type x: pandas.core.frame.DataFrame
+    :type x: pandas dataframe
     :param y: DV.
-    :type y: pandas.core.frame.DataFrame
+    :type y: pandas dataframe
     :param data: dataset.
-    :type data: pandas.core.frame.DataFrame
+    :type data: pandas dataframe
     :param weights: weights.
     :type weights: int
     :param offsetx: offset for X.
@@ -238,16 +238,16 @@ def op(pstart, x, y, data, weights, offsetx):
 def ziop(pstart, x, y, z, data, weights, offsetx, offsetz):
     """Calculate likelihood function for Zero-inflated Model.
 
-    :param pstart: starting parameters.
-    :type pstart: numpy.ndarray or list
+    :param pstart: A list of starting parameters.
+    :type pstart: list
     :param x: Ordered stage variables.
-    :type x: pandas.core.frame.DataFrame
+    :type x: pandas dataframe
     :param y: Dependent Variable (DV).
-    :type y: pandas.core.frame.DataFrame
+    :type y: pandas dataframe
     :param z: Inflation stage variables.
-    :type z: pandas.core.frame.DataFrame
+    :type z: pandas dataframe
     :param data: dataset.
-    :type data: pandas.core.frame.DataFrame
+    :type data: pandas dataframe
     :param weights: weights.
     :type weights: int
     :param offsetx: offset for X.
@@ -293,16 +293,16 @@ def ziop(pstart, x, y, z, data, weights, offsetx, offsetz):
 def ziopc(pstart, x, y, z, data, weights, offsetx, offsetz):
     """Calculate likelihood function for Zero-inflated Correlated-Errors Model.
 
-    :param pstart: starting parameters.
-    :type pstart: numpy.ndarray or list
+    :param pstart: A list of starting parameters.
+    :type pstart: list
     :param x: Ordered stage variables.
-    :type x: pandas.core.frame.DataFrame
+    :type x: pandas dataframe
     :param y: Dependent Variable (DV)
-    :type y: pandas.core.frame.DataFrame
+    :type y: pandas dataframe
     :param z: Inflation stage variables.
-    :type z: pandas.core.frame.DataFrame
+    :type z: pandas dataframe
     :param data: dataset.
-    :type data: pandas.core.frame.DataFrame
+    :type data: pandas dataframe
     :param weights: weights.
     :type weights: int
     :param offsetx: offset for X.
@@ -373,16 +373,16 @@ def miop(pstart, x, y, z, data, weights, offsetx, offsetz):
     "without" correlated errors.
     Number of outcomes must be odd.
 
-    :param pstart: starting parameters.
-    :type pstart: numpy.ndarray or list
+    :param pstart: A list of starting parameters.
+    :type pstart: list
     :param x: Ordered stage variables.
-    :type x: pandas.core.frame.DataFrame
+    :type x: pandas dataframe
     :param y: Dependent Variable (DV)
-    :type y: pandas.core.frame.DataFrame
+    :type y: pandas dataframe
     :param z: Inflation stage variables.
-    :type z: pandas.core.frame.DataFrame
+    :type z: pandas dataframe
     :param data: dataset.
-    :type data: pandas.core.frame.DataFrame
+    :type data: pandas dataframe
     :param weights: weights.
     :type weights: int
[…]
