### 2021-06-16 9ba288c
edit more docstrings
