### 2021-06-16 23e1000
more docstring edits
