+++ idcempy/gimnl.py
@@ -44,7 +44,7 @@ class GimnlModel:
             variables and missing values listwise deleted.
         :param zs: Inflation stage estimates (Gammas).
         :param xs: Multinomial Logit probit estimates (Betas).
-        :param ycatu: Number of categories in the Dependent Variable (DV).
+        :param ycatu: Number of categories in the Dependent variable (DV).
         :param x_: X Data.
         :param yx_: Y (DV) data.
         :param z_: Z Data.
@@ -105,7 +105,7 @@ class MnlModel:
         :param data: Model data used for estimation, subsetted to selected
             variables and missing values listwise deleted.
         :param xs: Multinomial Logit estimates (Betas).
-        :param ycatu: Number of categories in the Dependent Variable (DV).
+        :param ycatu: Number of categories in the Dependent variable (DV).
         :param x_: X Data.
         :param yx_: Y (DV) data.
         :param xstr: list of string for x names.
@@ -133,12 +133,12 @@ def mnl3(pstart, x2, x3, y, reference):
 
     :param pstart: A list of starting parameters.
     :type pstart: list
-    :param x2: Multinomial Logit covariates. Data subsetted to selected
+    :param x2: Multinomial Logit variables. Data subsetted to selected
         variables.
     :type x2: pandas dataframe
-    :param x3: Multinomial Logit covariates (should be identical to x2.
+    :param x3: Multinomial Logit variables (should be identical to x2.
     :type x3: pandas dataframe
-    :param y: Dependent Variable (DV). Data subsetted to selected
+    :param y: Dependent variable (DV). Data subsetted to selected
         variable.
     :type y: pandas dataframe
     :param reference: list of order of categories.
@@ -166,15 +166,15 @@ def bimnl3(pstart, x2, x3, y, z, reference):
 
     :param pstart: A list of starting parameters.
     :type pstart: list
-    :param x2: Multinomial Logit covariates. Data subsetted to selected
+    :param x2: Multinomial Logit variables. Data subsetted to selected
         variables.
     :type x2: pandas dataframe
-    :param x3: Multinomial Logit covariates (should be identical to x2.
+    :param x3: Multinomial Logit variables (should be identical to x2.
     :type x3: pandas dataframe
-    :param y: Dependent Variable (DV). Data subsetted to selected
+    :param y: Dependent variable (DV). Data subsetted to selected
         variable.
     :type y: pandas dataframe
-    :param z: Logit Split stage covariates. Data subsetted to selected
+    :param z: Logit Split stage variables. Data subsetted to selected
         variables.
     :type z: pandas dataframe
     :param reference: list of order of categories (first element will be the
@@ -206,15 +206,15 @@ def simnl3(pstart, x2, x3, y, z, reference):
 
     :param pstart: A list of starting parameters.
     :type pstart: list
-    :param x2: Multinomial Logit covariates. Data subsetted to selected
+    :param x2: Multinomial Logit variables. Data subsetted to selected
         variables.
     :type x2: pandas dataframe
-    :param x3: Multinomial Logit covariates (should be identical to x2.
+    :param x3: Multinomial Logit variables (should be identical to x2.
     :type x3: pandas dataframe
-    :param y: Dependent Variable (DV). Data subsetted to selected
+    :param y: Dependent variable (DV). Data subsetted to selected
         variable.
     :type y: pandas dataframe
-    :param z: Logit Split stage covariates. Data subsetted to selected
+    :param z: Logit Split stage variables. Data subsetted to selected
         variables.
     :type z: pandas dataframe
     :param reference: list of order of categories (second element will be the
@@ -246,15 +246,15 @@ def timnl3(pstart, x2, x3, y, z, reference):
 
     :param pstart: A list of starting parameters.
     :type pstart: list
[…]
