### 2021-06-17 ac9b6ba
Update gimnl_examples.py
