### 2021-06-18 4241f69
edits to gimnl tests
