### 2021-06-18 5a5009c
added unittest for zmiopc
