### 2021-06-19 f9b34c3
add tests for miop, change places to 0 for less restrictive assertions
