+++ docs/idcempy_tutorial.rst
@@ -39,7 +39,6 @@ The package can be installed in two different ways.
 1. From `PyPi <https://pypi.org/>`__:
 
 .. testcode::
-
   # Import the package
 
   pip install idcempy
@@ -47,7 +46,6 @@ The package can be installed in two different ways.
 2. From its `GitHub Repository <https://github.com/hknd23/idcempy/>`__:
 
 .. testcode::
-
   # Import the package
 
   git clone https://github.com/hknd23/idcempy.git
@@ -64,9 +62,9 @@ The `iopcmod` function estimates regression objects for "zero-inflated" and "mid
 We first import the required libraries, set up the package and import the dataset:
 
 .. testcode::
-
   # Import the necessary libraries and package
 
+  import numpy as np
   import pandas as pd
   import urllib
   from idcempy import zmiopc
@@ -76,6 +74,7 @@ We first import the required libraries, set up the package and import the datase
   url='https://github.com/hknd23/zmiopc/blob/main/data/tobacco_cons.csv'
 
   # Define a `Pandas` DataFrame
+  
   data = pd.read_csv(url)
 
 Our data is now a `pandas` DataFrame, and we can proceed to estimate the ZiOP model as follows.
@@ -98,8 +97,20 @@ The package sets a default start value of .01 for all parameters.  Users can mod
 .. testcode::
 
    # Model estimation:
-  ziop_tob= zmiopc.iopmod('ziop', data, X, Y, Z, method='bfgs', weights= 1,offsetx= 0, offsetz=0)
-
+   
+   ziop_tob= zmiopc.iopmod('ziop', data, X, Y, Z, method = 'bfgs', weights = 1, offsetx = 0, offsetz = 0)
+   
+   # 'ziop' = model to be estimated. In this case 'ziop'
+   # data = name of Pandas DataFrame 
+   # X = variables in the ordered probit stage.  
+   # Y = pependent variable.
+   # Z = variables in the inflation stage.
+   # method = method for optimization.  By default set to 'bfgs'
+   # weights = weights. 
+   # offsetx = offset of X.  By Default is zero.
+   # offsetz = offset of z
+   
+   
 Results from the model:
 
 The following message will appear when the model has converged:
@@ -159,6 +170,9 @@ You can also extract predicted probabilities from the model:
 .. testcode::
 
   fittedziop = ziopc.iopfit(ziop_tob)
+  
+  # Print the predicted probabilities
+  
   print(fittedziopc.responsefull)
 
 .. testoutput::
@@ -173,7 +187,7 @@ You can also extract predicted probabilities from the model:
 
 You can compute changes in predicted probabilities when the value of a variable changes.
 This allows you to illustrate how changes in the split-probit covariates affect the probabilities of
-being in one population versus another. The example below illustrates the marginal effects of the variable
+being in one population versus another. The example below, for instance,  illustrates the marginal effects of the variable
 'gender_dum' on the outcome variable in the ZiOP model estimated above.
 
 .. testcode::
@@ -187,6 +201,9 @@ You can also calculate the change in predicted probabilities of the outcome vari
 .. testcode::
 
     gender = zmiopc.ordered_effects(ziop_tob, 2, nsims = 10000)
+    
+    # You can now plot the predicted probabilities estimated earlier
+    
     gender.plot.box(grid='False')
 
 Zero-inflated Ordered Probit (ZiOPC) with Correlated Errors
@@ -196,9 +213,9 @@ The package also includes the function `iopcmod` which fits "zero-inflated" orde
 We first import the required libraries, set up the package and import the dataset:
 
 .. testcode::
-
[…]
