### 2021-06-19 e5c53fb
Update idcempy_tutorial.rst
