+++ docs/idcempy_tutorial.rst
@@ -5,7 +5,7 @@ IDCeMPy Package
 Description
 ===========
 `IDCeMPy` is a Python package that provides functions to fit and assess the performance of the following distinct
-sets of “inflated” discrete choice models.
+sets of “inflated” discrete choice models:
 
 * Fit the Zero-Inflated Ordered Probit (ZIOP) model without and with correlated errors (ZIOPC model) to evaluate zero-inflated ordered choice outcomes that results from a dual data generating process (d.g.p.).
 
@@ -20,7 +20,7 @@ sets of “inflated” discrete choice models.
 When should you use `IDCeMPy`?
 ==============================
 
-An excessive (“inflated”) share of observations—stemming from two distinct d.g.p’s—fall into a single choice category in many ordered and unordered polytomous outcome variables. Standard Ordered Probit and Multinomial Logit models cannot account for such category inflation which leads to biased inferences. Examples include,
+An excessive (“inflated”) share of observations—stemming from two distinct d.g.p’s—fall into a single choice category in many ordered and unordered polytomous outcome variables. Standard Ordered Probit and Multinomial Logit models cannot account for such category inflation which leads to biased inferences. Examples for such d.g.p’s include:
 
 * The inflated zero-category of "no smoking" in ordered measures of self-reported smoking behavior is generated from nonsmokers who never smoke cigarettes and those who smoked previously but temporarily stopped smoking because of high cigarette prices.
 
@@ -34,32 +34,29 @@ Each inflated discrete choice model in this package addresses category inflation
 
 Installation
 =============
-The package can be installed in two different ways.
+The package can be installed in two different ways:
 
-1. From `PyPi <https://pypi.org/>`__:
+1. From `PyPi <https://pypi.org/project/idcempy/>`__:
 
 .. testcode::
 
-  # Import the package
-
-  pip install idcempy
+  $  pip install idcempy
 
 2. From its `GitHub Repository <https://github.com/hknd23/idcempy/>`__:
 
 .. testcode::
 
-  # Import the package
-
-  git clone https://github.com/hknd23/idcempy.git
-  cd idcempy
-  python setup.py install
+  $  git clone https://github.com/hknd23/idcempy.git
+  $  cd idcempy
+  $  python setup.py install
 
 Examples
 ========
 
 Zero-inflated Ordered Probit (ZiOP) Model without Correlated Errors
 --------------------------------------------------------------------
-The `iopcmod` function estimates regression objects for "zero-inflated" and "middle-inflated" ordered probit models without correlated errors.  Below you will find instructions to estimate a ZiOP model.
+The `iopcmod` function estimates regression objects for "zero-inflated" and "middle-inflated" ordered probit models without correlated errors.
+This section provides instruction to estimate the ZiOP model using the self-reported smoking behavior as empirical example.
 
 We first import the required libraries, set up the package and import the dataset:
 
@@ -71,14 +68,12 @@ We first import the required libraries, set up the package and import the datase
   import urllib
   from idcempy import zmiopc
 
-  # Import the "Youth Tobacco Consumption" dataset.
+  # Import the "Youth Tobacco Consumption" dataset as a pandas.DataFrame
 
   url='https://github.com/hknd23/zmiopc/blob/main/data/tobacco_cons.csv'
-
-  # Define a `Pandas` DataFrame
   data = pd.read_csv(url)
 
-Our data is now a `pandas` DataFrame, and we can proceed to estimate the ZiOP model as follows.
+The data is now a `pandas` DataFrame, and we can proceed to estimate the ZiOP model as follows.
 
 .. testcode::
 
[…]
