### 2021-06-19 f65f2cb
remove "gtol": 1e-6,  in minimization for more flexible adaption to other methods
