+++ docs/idcempy_tutorial.rst
@@ -207,7 +207,7 @@ Results from :func:`zmiopc.split_effects` and :func:`zmiopc.ordered_effects` can
 
 Zero-inflated Ordered Probit (ZiOPC) with Correlated Errors
 -----------------------------------------------------------
-The package also includes the function `iopcmod` which fits "zero-inflated" ordered probit models (ZiOPC) under the assumption that the two errors are correlated with each other (i.e. correlated errors).
+The package also includes :func:`zmiopc.iopcmod` which fits "zero-inflated" ordered probit models (ZiOPC) under the assumption that the two errors are correlated with each other (i.e. correlated errors).
 
 We first import the required libraries, set up the package and import the dataset:
 
@@ -237,7 +237,7 @@ We first import the required libraries, set up the package and import the datase
   Z = ['gender_dum']
   Y = ['cig_count']
 
-Our data is now a `Pandas` DataFrame, and we can proceed to estimate the ZiOP model as follows.
+:func:`zmiopc.iopcmod` estimates the ZiOPC model using the keyword `'ziopc'` in the first argument:
 
 .. testcode::
 
@@ -253,9 +253,6 @@ Our data is now a `Pandas` DataFrame, and we can proceed to estimate the ZiOP mo
    # offsetx = offset of X.  By Default is zero.
    # offsetz = offset of z
 
-
-The package sets a default start value of .01 for all parameters.  Users can modify it by creating an array with their desired values, define such array as `pstart` and add it to as an argument in the model function.
-
 The results are stored in the attributes of :class:`zmiopc.IopCModel`.
 
 .. testoutput::
@@ -265,7 +262,7 @@ The results are stored in the attributes of :class:`zmiopc.IopCModel`.
          Function evaluations: 1562
          Gradient evaluations: 142
 
-The following line of code prints the results
+The following line of code prints the results:
 
 .. testcode::
 
@@ -285,7 +282,7 @@ The following line of code prints the results
    Ordered: gender_dum    0.602136  0.053084  11.343020  0.000000e+00   0.498091   0.706182
    rho                   -0.415770  0.074105  -5.610526  2.017123e-08  -0.561017  -0.270524
 
-To print the estimates of the log-likelihood, AIC, and Variance-Covariance matrix, you should type:
+To print the estimates of the log-likelihood, AIC, and Variance-Covariance matrix:
 
 .. testcode::
 
@@ -307,8 +304,7 @@ The AIC of the ziopc_tob model, for example, is:
 
   10140.103819465658
 
-The predicted probabilities from the `ziopc_tob` model can ve obtained as follows.
-
+The predicted probabilities from the `ziopc_tob` model can be obtained with :func:`zmiopc.iopcfit` as follows.
 
 .. testcode::
 
@@ -330,16 +326,12 @@ The predicted probabilities from the `ziopc_tob` model can ve obtained as follow
  [0.87523652 0.06888286 0.01564958 0.0275354  0.01269564]
  [0.82678185 0.0875059  0.02171135 0.04135142 0.02264948]]
 
-You can compute changes in predicted probabilities when the value of a variable changes.
-This allows you to illustrate how the changes in the split-probit covariates affect the probabilities of being in one population versus another. The example below illustrates the marginal effects of the variable 'gender_dum' on the outcome variable in the ZiOPC model estimated in ths documentation.
+ Similar to the ZiOP model, :func:`zmiopc.split_effects` and :func:`zmiopc.ordered_effects` can also
+ compute changes in predicted probabilities for the ZiOPC model.
 
 .. testcode::
 
-    ziopcgender = idcempy.split_effects(ziopc_tob, 1, nsims = 10000)
-
[…]
