+++ docs/idcempy_tutorial.rst
@@ -100,7 +100,7 @@ The package sets a default start value of .01 for all parameters.
    # 'ziop' = model to be estimated. In this case 'ziop'
    # data = name of Pandas DataFrame
    # X = variables in the ordered probit stage.
-   # Y = pependent variable.
+   # Y = dependent variable.
    # Z = variables in the inflation stage.
    # method = method for optimization.  By default set to 'bfgs'
    # weights = weights.
@@ -207,6 +207,7 @@ Results from :func:`zmiopc.split_effects` and :func:`zmiopc.ordered_effects` can
 
 Zero-inflated Ordered Probit (ZiOPC) with Correlated Errors
 -----------------------------------------------------------
+
 The package also includes :func:`zmiopc.iopcmod` which fits "zero-inflated" ordered probit models (ZiOPC) under the assumption that the two errors are correlated with each other (i.e. correlated errors).
 
 We first import the required libraries, set up the package and import the dataset:
@@ -246,7 +247,7 @@ We first import the required libraries, set up the package and import the datase
    # 'ziopc' = model to be estimated. In this case 'ziopc'
    # data = name of Pandas DataFrame
    # X = variables in the ordered probit stage.
-   # Y = pependent variable.
+   # Y = dependent variable.
    # Z = variables in the inflation stage.
    # method = method for optimization.  By default set to 'bfgs'
    # weights = weights.
@@ -344,6 +345,7 @@ The predicted probabilities from the `ziopc_tob` model can be obtained with :fun
 
 Middle-inflated Ordered Probit (MiOP) without Correlated Errors
 ---------------------------------------------------------------
+
 A Middle-inflated Ordered Probit (MiOP) model should be estimated ehn the ordered outcome variable is inflated in the middle category.
 
 The following example uses 2004 presidential vote data from Elgun and Tilam (`2007 <https://journals.sagepub.com/doi/10.1177/1065912907305684>`_).
@@ -390,7 +392,7 @@ After importing the dataset and specifying the list of variables from it, the Mi
    # 'miop' = Type of model to be estimated. In this case 'miop'
    # data = name of Pandas DataFrame
    # X = variables in the ordered probit stage.
-   # Y = pependent variable.
+   # Y = dependent variable.
    # Z = variables in the inflation stage.
    # method = method for optimization.  By default set to 'bfgs'
    # weights = weights.
@@ -532,7 +534,7 @@ The model can be estimated as follows:
    # 'miopc' = Type of model to be estimated. In this case 'miopc'
    # data = name of Pandas DataFrame
    # X = variables in the ordered probit stage.
-   # Y = pependent variable.
+   # Y = dependent variable.
    # Z = variables in the inflation stage.
    # method = method for optimization.  By default set to 'BFGS'
    # weights = weights.
@@ -587,14 +589,13 @@ To calculate the predicted probabilities:
 
    print(fittedziopc.responsefull)
 
-The following line of code allows you to compute changes in predicted probabilities when the value of a variable changes.
-This allows you to illustrate how the changes in the split-probit covariates affect the probablilities of being in one population versus another.
+The following line of code computes changes in predicted probabilities when the value of a variable changes.
 
 .. testcode::
 
    # Define model from which effects will be estimated and number of simulations
 
-   miopcxeno = idcempy.split_effects(miopc_EU, 1, nsims = 10000)
+   miopcxeno = zmiopc.split_effects(miopc_EU, 1, nsims = 10000)
 
 A box plot can illustrate the change in predicted probabilities.
 
@@ -621,10 +622,10 @@ To calculate the change in predicted probabilities of the outcome variable in th
 The Standard Ordered Probit (OP) model
 --------------------------------------
 
[…]
