### 2021-06-20 a276bb0
Update idcempy_tutorial.rst
