### 2021-06-20 bc79c28
Update idcempy_tutorial.rst
