+++ idcempy/zmiopc.py
@@ -27,8 +27,8 @@ class OpModel:
         :param yncat: Number of categories in the outcome variable.
         :param x_: X (Covariates) Data.
         :param yx_: Y (Dependent Variable) data.
-        :param xstr: list of strings for variable(s) in the outcome stage (x).
-        :param ystr: list of strings for outcome variable name (y).
+        :param xstr: List of strings for variable(s) in the outcome stage (x).
+        :param ystr: List of strings for outcome variable name (y).
         """
         self.llik = llik
         self.coefs = coef
@@ -82,9 +82,9 @@ class IopModel:
         :param x_: X Data.
         :param yx_: Y (DV) data.
         :param z_: Z Data.
-        :param xstr: list of strings for x names.
-        :param ystr: list of strings for y names.
-        :param zstr: list of strings for z names.
+        :param xstr: List of strings for x names.
+        :param ystr: List of strings for y names.
+        :param zstr: List of strings for z names.
 
         """
         self.modeltype = modeltype
@@ -145,9 +145,9 @@ class IopCModel:
         :param x_: X Data.
         :param yx_: Y (DV) data.
         :param z_: Z Data.
-        :param xstr: list of strings for x names.
-        :param ystr: list of strings for y names.
-        :param zstr: list of strings for z names.
+        :param xstr: List of strings for x names.
+        :param ystr: List of strings for y names.
+        :param zstr: List of strings for z names.
         """
         self.modeltype = modeltype
         self.llik = llik
@@ -189,17 +189,17 @@ class FittedVals:
 def op(pstart, x, y, data, weights, offsetx):
     """Calculate likelihood function for Ordered Probit Model.
 
-    :param pstart: A list of starting parameters.
+    :param pstart: A list of starting values for the estimation.  Length of the number of parameters. 
     :type pstart: list
-    :param x: Ordered stage variables.
+    :param x: Covariates for the ordered stage.
     :type x: pandas dataframe
-    :param y: DV.
+    :param y: The dependent variable (DV)
     :type y: pandas dataframe
     :param data: dataset.
     :type data: pandas dataframe
     :param weights: weights.
     :type weights: int
-    :param offsetx: offset for X.
+    :param offsetx: Offset for covariates in the ordered stage.
     :type offsetx: int
     """
     n = len(data)
@@ -238,24 +238,21 @@ def op(pstart, x, y, data, weights, offsetx):
 def ziop(pstart, x, y, z, data, weights, offsetx, offsetz):
     """Calculate likelihood function for Zero-inflated Model.
 
-    :param pstart: A list of starting parameters.
+    :param pstart: A list of starting values for the estimation.  Length of the number of parameters. 
     :type pstart: list
-    :param x: Ordered stage variables. Data subsetted to selected
-        variables.
+    :param x: Covariates for the ordered probit stage. 
     :type x: pandas dataframe
-    :param y: Dependent Variable (DV). Data subsetted to selected
-        variables.
+    :param y: The ordinal dependent Variable (DV). 
     :type y: pandas dataframe
-    :param z: Inflation stage variables. Data subsetted to selected
-        variables.
+    :param z: Covariates for the inflation stage. 
     :type z: pandas dataframe
     :param data: Dataset with missing values listwise deleted.
     :type data: pandas dataframe
     :param weights: weights.
     :type weights: float
-    :param offsetx: offset for X.
+    :param offsetx: Offset for the ordered stage covariates (X). 
     :type offsetx: float
-    :param offsetz: offset for z.
+    :param offsetz: Offset for the inflation stage covariates (Z).
     :type offsetz: float
     """
     n = len(data)
@@ -296,24 +293,21 @@ def ziop(pstart, x, y, z, data, weights, offsetx, offsetz):
[…]
