+++ paper.bib
@@ -73,13 +73,16 @@
   publisher={SAGE Publications Sage CA: Los Angeles, CA}
 }
 
-@article{dale2018estimation,
-  title={Estimation of Nested and Zero-Inflated Ordered Probit Models},
+@article{dale2021estimation,
+  title={Estimation of nested and zero-inflated ordered probit models},
   author={Dale, David and Sirchenko, Andrei},
-  journal={Higher School of Economics Research Paper No. WP BRP},
-  volume={193},
-  year={2018},
-  doi = {10.2139/ssrn.3187203}
+  journal={The Stata Journal},
+  volume={21},
+  number={1},
+  pages={3--38},
+  year={2021},
+  publisher={SAGE Publications Sage CA: Los Angeles, CA},
+  doi = {10.1177/1536867X211000002}
 }
 
 @article{bagozzi2017distinguishing,
@@ -184,3 +187,102 @@
   publisher={Cambridge University Press},
   doi = {10.1017/CBO9780511845062}
 }
+
+@article{sarrias2016discrete,
+  title={Discrete choice models with random parameters in R: The Rchoice Package},
+  author={Sarrias, Mauricio and others},
+  journal={Journal of Statistical Software},
+  volume={74},
+  number={10},
+  pages={1--31},
+  year={2016},
+  publisher={Foundation for Open Access Statistics},
+  doi = {10.18637/jss.v074.i10}
+  }
+  
+  @incollection{richards2018new,
+  title={New empirical models in consumer demand},
+  author={Richards, Timothy J and Bonnet, Celine},
+  booktitle={The Routledge Handbook of Agricultural Economics},
+  pages={488--511},
+  year={2018},
+  publisher={Routledge},
+  doi = {10.4324/9781315623351-27}
+  }
+  
+  @article{brown2020modelling,
+  title={Modelling Category Inflation with Multiple Inflation Processes: Estimation, Specification and Testing 1},
+  author={Brown, Sarah and Harris, Mark N and Spencer, Christopher},
+  journal={Oxford Bulletin of Economics and Statistics},
+  volume={82},
+  number={6},
+  pages={1342--1361},
+  year={2020},
+  publisher={Wiley Online Library},
+  doi = {10.1111/obes.12366}  
+  }
+  
+@incollection{venables2002random,
+  title={Random and mixed effects},
+  author={Venables, WN and Ripley, BD},
+  booktitle={Modern applied statistics with S},
+  pages={271--300},
+  year={2002},
+  publisher={Springer},
+  doi = {10.1007/978-0-387-21706-2}
+  }
+  
+  @article{henningsen2014micecon,
+  title={micEcon: Microeconomic Analysis and Modelling},
+  author={Henningsen, A},
+  journal={R package version 0.6-12, URL http://CRAN. R-project. org/package= micEcon},
+  year={2014}
+}
+
+@article{croissant2012estimation,
+  title={Estimation of multinomial logit models in R: The mlogit Packages},
+  author={Croissant, Yves and others},
+  journal={R package version 0.2-2. URL: http://cran. r-project. org/web/packages/mlogit/vignettes/mlogit. pdf},
+  year={2012},
+  publisher={Citeseer}
+}
+
+@article{hasan2014fast,
+  title={Fast estimation of multinomial logit models: R package mnlogit},
+  author={Hasan, Asad and Zhiyu, Wang and Mahani, Alireza S},
+  journal={arXiv preprint arXiv:1404.3177},
+  year={2014}
+}
+
+@article{sarrias2015gmnl,
+  title={gmnl: Multinomial Logit Models with Random Parameters},
+  author={Sarrias, M and Daziano, R},
[…]
