+++ docs/idcempy_tutorial.rst
@@ -17,7 +17,7 @@ sets of “inflated” discrete choice models:
 
 `IDCeMPy` uses Newton numerical optimization methods to estimate the models listed above via Maximum Likelihood Estimation (MLE).
 
-When should you use `IDCeMPy`?
+When Should You use `IDCeMPy`?
 ==============================
 
 An excessive (“inflated”) share of observations—stemming from two distinct d.g.p’s—fall into a single choice category in many ordered and unordered polytomous outcome variables. Standard Ordered Probit and Multinomial Logit models cannot account for such category inflation which leads to biased inferences. Examples for such d.g.p’s include:
@@ -64,7 +64,6 @@ We first import the required libraries, set up the package and import the datase
 
   # Import the necessary libraries and package
 
-  import numpy as np
   import pandas as pd
   import urllib
   from idcempy import zmiopc
@@ -78,7 +77,7 @@ The data is now a `pandas` DataFrame, and we can proceed to estimate the ZiOP mo
 
 .. testcode::
 
-  # First, you should define a list of variable names of X, Z, and Y.
+  # First, define a list of variable names of X, Z, and Y.
   # X = Column names of covariates (from `DataFrame`) used in ordered probit stage.
   # Z = Column names of covariates (from `DataFrame`) used in split-population stage.
   # Y = Column name of ordinal outcome variable (from `DataFrame`).
@@ -218,7 +217,6 @@ We first import the required libraries, set up the package and import the datase
 
   # Import the necessary libraries and IDCeMPy.
 
-  import numpy as np
   import pandas as pd
   import urllib
   from idcempy import zmiopc
@@ -232,7 +230,7 @@ We first import the required libraries, set up the package and import the datase
 
 .. testcode::
 
-  # First, you should define a list of variable names of X, Z, and Y.
+  # First, define a list of variable names of X, Z, and Y.
   # X = Column names of covariates (from `DataFrame`) used in ordered probit stage.
   # Z = Column names of covariates (from `DataFrame`) used in split-population stage.
   # Y = Column name of ordinal outcome variable (from `DataFrame`).
@@ -380,7 +378,7 @@ We then define the lists with the names of the variables used in the model
 
 .. testcode::
 
-  # First, you should define a list of variable names of X, Z, and Y.
+  # First, define a list of variable names of X, Z, and Y.
   # X = Column names of covariates (from `DataFrame`) used in ordered probit stage.
   # Z = Column names of covariates (from `DataFrame`) used in split-population stage.
   # Y = Column name of ordinal outcome variable (from `DataFrame`).
@@ -504,7 +502,6 @@ First is importing the data and libraries:
 
   # Import the necessary libraries and IDCeMPy.
 
-  import numpy as np
   import pandas as pd
   import urllib
   from idcempy import zmiopc
@@ -525,7 +522,7 @@ We then define the lists with the names of the variables used in the model:
 
 .. testcode::
 
-   # First, you should define a list of variable names of X, Z, and Y.
+   # First, define a list of variable names of X, Z, and Y.
    # X = Column names of covariates (from `DataFrame`) used in ordered probit stage.
    # Z = Column names of covariates (from `DataFrame`) used in split-population stage.
    # Y = Column name of ordinal outcome variable (from `DataFrame`).
@@ -643,7 +640,6 @@ First, import the required libraries and data:
 
    # Import the necessary libraries and package
 
-   import numpy as np
    import pandas as pd
    import urllib
    from idcempy import zmiopc
@@ -763,7 +759,7 @@ The Vuong test can also be implemented to compare the ZiOPC, MiOP and MiOPC mode
 Generalized Inflated Multinomial Logit (GiMNL) Model
 ----------------------------------------------------
 
[…]
