### 2021-06-20 8bd0882
update_vuong_formula
