+++ paper.bib
@@ -10,6 +10,13 @@
   doi = {10.1162/ajhe_a_00026}
 }
 
+@book{greene2002nlogit,
+  title={NLOGIT: Version 3.0; Reference Guide},
+  author={Greene, William H},
+  year={2002},
+  publisher={Econometric Software, Incorporated}
+}
+
 @article{harris2007zero,
   title={A zero-inflated ordered probit model, with an application to modelling tobacco consumption},
   author={Harris, Mark N and Zhao, Xueyan},
