### 2021-06-20 350bf58
add llike functions to unittest
