### 2021-06-20 41977f5
Create CONTRIBUTING.md
