+++ docs/CONTRIBUTING.md
@@ -1,0 +1,8 @@
+# How to request support
+
+Questions about IDCeMPy can be sent to <nkh8@psu.edu> or by opening a new issue. 
+
+# How to contribute
+
+We welcome contributors to IDCeMPy. Issues can be raised by any user to report bugs. 
+For enhancement suggestions such as improvements to functionality or new features, please email <nkh8@psu.edu> for prompt response.
