+++ paper.md
@@ -12,7 +12,7 @@ authors:
   name: Vineeta Yadav
 - affiliation: 1
   name: Bumba Mukherjee
-date: "16 February 2021"
+date: "21 June 2021"
 output:
   html_document:
     df_print: paged
@@ -29,125 +29,31 @@ affiliations:
   name: Dept. of Political Science, San Jose State University
   
 ---
-# Summary
-`IDCeMPy` is a free, open-source and versatile Python package that enables
-researchers to easily fit three distinct sets of inflated discrete choice models that
-are often used in a variety of disciplines including Economics, Business, Engineering, Political
-Science, Psychology, and Public Health: (i) Zero-Inflated Ordered Probit (ZIOP), (ii)
-Middle-Inflated Ordered Probit (MIOP), and (iii) Generalized Inflated Multinomial
-Logit (GIMNL) models. While the ZIOP model [@harris2007zero] permits careful evaluation 
-of zero-inflated ordered choice outcomes that results from two data generating
-processes (hereafter, “d.g.p’s”), the MIOP model [@bagozzi2012mixture]
-accounts for ordered choice outcomes in which the inflated middle-category emerges 
-from distinct d.g.p’s. The GIMNL models [@bagozzi2017distinguishing] account for the large proportion—and heterogeneous
-mixture—of observations in the baseline and other unordered outcome
-categories within MNL models that evaluate multiple unordered polytomous
-outcomes.
+#Summary
 
-`IDCeMPy` is currently intended to be used by researchers in social, behavioral,
-and natural sciences, and by data scientists working with inflated discrete choice
-outcome variables. Its clean, easy to use Python code is well-tested and well-
-documented. For further details, the reader is referred to the official `IDCeMPy`
-repository.
+Scholars and Data Scientists often use discrete choice models to evaluate ordered dependent variables using the ordered probit model and unordered polytomous outcome measures via the multinomial logit (MNL) estimator [@greene2002nlogit; @sarrias2016discrete; @richards2018new]. These models, however, cannot account for the possibility that in many ordered and unordered polytomous choice outcomes, an excessive share of observations —stemming from two distinct data generating processes (d.g.p’s)— fall into a single category which is thus “inflated.” For instance, ordered outcome measures of self-reported smoking behavior that range from 0 for “no smoking” to 3 for “smoking 20 cigarettes or more daily” contain excessive observations in the zero (no smoking) category that includes individuals who never smoke cigarettes and those who smoked previously but temporarily stop smoking because of an increase in cigarette costs [@harris2007zero; @greene2015inflated]. The “indifference” middle-category in ordered measures of immigration attitudes is inflated since it includes respondents who are truly indifferent about immigration and those who select “indifference” because of social desirability reasons [@bagozzi2012mixture; @brown2020modelling]. The baseline category of unordered polytomous variables of Presidential vote choice is also often inflated as it includes non-voters who abstain from voting owing to temporary factors and routine non-voters who are disengaged from the political process [@campbell2008religion; @bagozzi2017distinguishing].  Inflated discrete choice models have been developed to address such category inflation in ordered and unordered polytomous outcome variables as failing to do so leads to model misspecification and incorrect inferences [@harris2007zero; @bagozzi2012mixture; @brown2020modelling].
 
-# Statement of Need
-Scholars and Data Scientists often use discrete choice models to evaluate
-ordered outcomes using the ordered probit model and unordered polytomous
-measures via the multinomial logit (MNL) estimator. These models, however,
-cannot account for the fact that in many ordered and unordered polytomous
-choice situations, an excessive share of observations—stemming from two
-distinct d.g.p’s—fall into a single category which is thus “inflated.” For instance,
[…]
