+++ paper.md
@@ -37,7 +37,7 @@ Scholars and Data Scientists often use discrete choice models to evaluate ordere
 
 # State of the Field
 
-Software packages and code are available for estimating standard (non-inflated) discrete choice models. In the R environment, the packages `MASS` [@venables2002random] and `micEcon` [@henningsen2014micecon] fit binary and discrete choice models. The package `Rchoice` [@sarrias2016discrete] allows researchers to estimate binary as well as ordered probit and logit models and the Poisson model by employing various optimization routines. The proprietary LIMDEP package NLOGIT [@greene2002nlogit] fits conventional binary and ordered discrete choice models but is neither open-sourced nor freely available. The R packages `mlogit` [@croissant2012estimation] and `mnlogit` [@hasan2014fast] provide tools for working with conventional MNL models, while `gmnl` [@sarrias2015gmnl] and `PReMiuM` [@liverani2015premium] estimates MNL models that incorporate unit-specific heterogeneity. There exists proprietary LIMDEP software and R code —but not an R package— that fit few inflated ordered probit and MNL models [@harris2007zero; @bagozzi2012mixture; @bagozzi2017distinguishing]. Outside R, the Python package `biogeme` [@bierlaire2016pythonbiogeme] fits mixed logit and MNL models. Further, @dale2018estimation’s ZiOP STATA command (but not package) fits the Zero-Inflated Ordered Probit without correlated errors. @xia2019gidm’s `gidm` STATA command fits discrete choice models without correlated errors for inflated zero and other lower-category discrete outcomes. 
+Software packages and code are available for estimating standard (non-inflated) discrete choice models. In the R environment, the packages `MASS` [@venables2002random] and `micEcon` [@henningsen2014micecon] fit binary and discrete choice models. The package `Rchoice` [@sarrias2016discrete] allows researchers to estimate binary as well as ordered probit and logit models and the Poisson model by employing various optimization routines. The proprietary LIMDEP package NLOGIT [@greene2002nlogit] fits conventional binary and ordered discrete choice models but is neither open-sourced nor freely available. The R packages `mlogit` [@croissant2012estimation] and `mnlogit` [@hasan2014fast] provide tools for working with conventional MNL models, while `gmnl` [@sarrias2015gmnl] and `PReMiuM` [@liverani2015premium] estimates MNL models that incorporate unit-specific heterogeneity. There exists proprietary LIMDEP software and R code —but not an R package— that fit few inflated ordered probit and MNL models [@harris2007zero; @bagozzi2012mixture; @bagozzi2017distinguishing]. Outside R, the Python package `biogeme` [@bierlaire2016pythonbiogeme] fits mixed logit and MNL models. Further, @dale2021estimation’s ZiOP STATA command (but not package) fits the Zero-Inflated Ordered Probit without correlated errors. @xia2019gidm’s `gidm` STATA command fits discrete choice models without correlated errors for inflated zero and other lower-category discrete outcomes. 
 
 The software written in R or LIMDEP and the STATA commands listed above are undoubtedly useful. Yet, to our knowledge, there does not exist any R or Python package to fit a variety of statistical models that account for the excessive (i.e., “inflated”) share of observations in the baseline and other higher categories of ordered and unordered polytomous dependent variables, which are commonly analyzed across the natural and social sciences. As discussed below, our Python package `IDCeMPy` thus fills an important lacuna by providing an array of functions that fit a substantial range of inflated discrete choice models that are applicable across various disciplines.
 
