+++ paper.bib
@@ -197,7 +197,7 @@
 
 @article{sarrias2016discrete,
   title={Discrete choice models with random parameters in R: The Rchoice Package},
-  author={Sarrias, Mauricio and others},
+  author={Sarrias, Mauricio},
   journal={Journal of Statistical Software},
   volume={74},
   number={10},
@@ -248,7 +248,7 @@
 
 @article{croissant2012estimation,
   title={Estimation of multinomial logit models in R: The mlogit Packages},
-  author={Croissant, Yves and others},
+  author={Croissant, Yves},
   journal={R package version 0.2-2. URL: http://cran. r-project. org/web/packages/mlogit/vignettes/mlogit. pdf},
   year={2012},
   publisher={Citeseer}
@@ -257,8 +257,12 @@
 @article{hasan2014fast,
   title={Fast estimation of multinomial logit models: R package mnlogit},
   author={Hasan, Asad and Zhiyu, Wang and Mahani, Alireza S},
-  journal={arXiv preprint arXiv:1404.3177},
-  year={2014}
+  journal={Journal of Statistical Software},
+  volume={75},
+  number={3},
+  pages={1--24},  
+  year={2016},
+  doi = {10.18637/jss.v075.i03 } 
 }
 
 @article{sarrias2015gmnl,
