+++ paper.bib
@@ -207,6 +207,17 @@
   doi = {10.18637/jss.v074.i10}
   }
   
+  @article{JSSv074i10,
+   author = {Mauricio Sarrias},
+   title = {Discrete Choice Models with Random Parameters in R: The Rchoice Package},
+   journal = {Journal of Statistical Software, Articles},
+   volume = {74},
+   number = {10},
+   year = {2016},
+   pages = {1--31},
+   doi = {10.18637/jss.v074.i10}
+}   
+   
   @incollection{richards2018new,
   title={New empirical models in consumer demand},
   author={Richards, Timothy J and Bonnet, Celine},
