+++ paper.bib
@@ -208,8 +208,8 @@
   }
   
   @article{JSSv074i10,
-   author = {Mauricio Sarrias},
    title = {Discrete Choice Models with Random Parameters in R: The Rchoice Package},
+   author = {Sarrias, Mauricio},
    journal = {Journal of Statistical Software, Articles},
    volume = {74},
    number = {10},
