### 2021-06-21 84a9b55
Update idcempy_tutorial.rst
