+++ docs/contact.rst
@@ -1,8 +1,4 @@
 Contact
 =======
 
-For more information, please contact nkh8@psu.edu
-
-R version of the package is available at: 
-
-
+For more information, please contact nkh8@psu.edu. 
