+++ paper.md
@@ -33,7 +33,7 @@ affiliations:
 
 Scholars and Data Scientists often use discrete choice models to evaluate ordered dependent variables using the ordered probit model and unordered polytomous outcome measures via the multinomial logit (MNL) estimator [@greene2002nlogit; @JSSv074i10; @richards2018new]. These models, however, cannot account for the possibility that in many ordered and unordered polytomous choice outcomes, an excessive share of observations —stemming from two distinct data generating processes (d.g.p’s)— fall into a single category which is thus “inflated.” For instance, ordered outcome measures of self-reported smoking behavior that range from 0 for “no smoking” to 3 for “smoking 20 cigarettes or more daily” contain excessive observations in the zero (no smoking) category that includes individuals who never smoke cigarettes and those who smoked previously but temporarily stop smoking because of an increase in cigarette costs [@harris2007zero; @greene2015inflated]. The “indifference” middle-category in ordered measures of immigration attitudes is inflated since it includes respondents who are truly indifferent about immigration and those who select “indifference” because of social desirability reasons [@bagozzi2012mixture; @brown2020modelling]. The baseline category of unordered polytomous variables of Presidential vote choice is also often inflated as it includes non-voters who abstain from voting owing to temporary factors and routine non-voters who are disengaged from the political process [@campbell2008religion; @bagozzi2017distinguishing].  Inflated discrete choice models have been developed to address such category inflation in ordered and unordered polytomous outcome variables as failing to do so leads to model misspecification and incorrect inferences [@harris2007zero; @bagozzi2012mixture; @brown2020modelling].
 
-`IDCeMPy` is an open-source Python package that enables researchers to fit three distinct sets of discrete choice models that are used by data scientists, economists, engineers, political scientists and public health researchers: the Zero-Inflated Ordered Probit (ZiOP) model without and with correlated errors (ZiOPC model), Middle-Inflated Ordered Probit (MiOP) model without and with correlated errors (MiOPC), and Generalized-Inflated Multinomial Logit (GiMNL) models. Functions that fit the ZiOP(C) model in `IDCeMPy` evaluates zero-inflated ordered dependent variables that results from two d.g.p’s, while functions that fit the MiOP(C) models account for inflated middle-category ordered outcomes that emerge from distinct d.g.p’s. The functions in IDCeMPy that fit GIMNL models account for the large share and heterogenous mixture of observations in the baseline and other lower outcome categories in unordered polytomous dependent variables. The primary location for the description of the functions that fit the models listed above is available at the [IDCeMPy package’s documentation website](https://idcempy.readthedocs.io/en/latest/).
+`IDCeMPy` is an open-source Python package that enables researchers to fit three distinct sets of discrete choice models that are used by data scientists, economists, engineers, political scientists and public health researchers: the Zero-Inflated Ordered Probit (ZiOP) model without and with correlated errors (ZiOPC model), Middle-Inflated Ordered Probit (MiOP) model without and with correlated errors (MiOPC), and Generalized-Inflated Multinomial Logit (GiMNL) models. Functions that fit the ZiOP(C) model in `IDCeMPy` evaluates zero-inflated ordered dependent variables that results from two d.g.p’s, while functions that fit the MiOP(C) models account for inflated middle-category ordered outcomes that emerge from distinct d.g.p’s. The functions in IDCeMPy that fit GiMNL models account for the large share and heterogenous mixture of observations in the baseline and other lower outcome categories in unordered polytomous dependent variables. The primary location for the description of the functions that fit the models listed above is available at the [IDCeMPy package’s documentation website](https://idcempy.readthedocs.io/en/latest/).
 
 # State of the Field
 
