### 2021-06-22 hknd23/idcempy#13 @hknd23
We incorporated your suggestion by dropping the phrase “efficient time-wise” in our revised paper. We briefly mention, however, that the time taken by the code in our IDCEMPy package to estimate in our data (the run-time for) the inflated discrete choice models without correlated errors (ZiOP, MiOP, GiMNL) ranges from approximately 8 to 75 seconds, while the run-time for the correlated error models (ZiOPC, MiOPC) vary from about 5 to almost 90 minutes. The run-time from the R code for the inflated (i) non-correlated error models in the data range from 9 to 15 minutes, and (ii) correlated error models is roughly 15 to 163 minutes.
