### 2021-06-22 hknd23/idcempy#16 @hknd23
We have added the contributing file to the package’s GitHub repository as well as a section for community guidelines in the documentation. Issues can be raised for questions and bug reports, and contributors can contact the authors through email for suggestions.
