### 2021-06-22 hknd23/idcempy#17 @hknd23
Thank you for raising this crucial point. To address this point, we have added a “State of the Field” section in our revised paper that first extensively cites and provides a brief review (given space constraints) of existing proprietary or open-source software packages—primarily written in R—such as MASS, Rchoice, NLOGIT and mnlogit that provide code to estimate standard discrete choice models including the binary probit, ordered probit and MNL models. While these packages are undoubtedly useful, we note that they cannot be employed to analyze a whole host of discrete choice outcome variables that exhibit category inflation and which are used across many disciplines (e.g., Economics, Engineering, Political Science, Public Health, and Sociology). Further, as noted in an earlier response, we also cite existing STATA commands and R code that fit a small handful of baseline or zero-inflated ordered probit and MNL models. We then discuss in our paper’s “Statement of Need” section how our package IDCEMPy package builds on but also substantially departs from—and thus contributes to—existing proprietary and open-source software on parametric (inflated) discrete choice models.    Further, we also added the citation of the Pandas Python package in our paper.
