### 2021-06-22 hknd23/idcempy#18 @hknd23
We answer these key questions in our revised paper “State of the Field” section by first noting that there currently exists the following two publicly available STATA commands (not packages) that each fit a specific type of a baseline or zero-inflated discrete choice model without correlated errors: Dale and Sirchenko’s (2021) ZiOP command and Xia et al’s (2019) gidm command. We also mention in the section mentioned above that there exists R code—again, not an R package—that fits few inflated ordered probit and MNL models. Building on this, we emphasize in the paper’s “State of the Field” and “Statement of Need” sections that there does not exist any R or Python package to fit a variety of statistical models that account for the “inflated” share of observations in the baseline and other higher categories of ordered and unordered polytomous dependent variables. Further, as discussed in our response below to your query about scholarly effort, we explain in our revised Statement of Need section how our Python package IDCEMPy addresses the lacuna mentioned above by providing functions to fit several inflated discrete choice models that are applicable to assess various issue-areas across disciplines as diverse as Economics, Engineering, Marketing research, Political Science, Public Health, Sociology, and Transportation research.
