### 2021-06-22 #3322 @hknd23
Dear @cmaimone,

We thank you very much for providing incisive, highly constructive, and detailed feedback and comments. Your review helped us understand numerous shortcomings in our package, documentation, and paper. We thus implemented all the following suggestions you raised in our revised work: reorganize and edit plus clarify various parts of the documentation associated with our IDCeMPy package, conduct additional tests to assess whether the functions in our package execute without error, revise some sections of our paper to review the “state of the field” and emphasize our package’s contributions in this regard, and provide additional details in our API documentation to improve our package’s features and functionality. Incorporating these suggestions helped us ensure that the revised version of our package and paper far better presents the work than our initial submission. We have responded to every issue you have raised in our package repository.

Regarding your point on Scholarly Effort, we implemented three main sets of revisions in the paper, our IDCeMPy package, and the package’s documentation on GitHub to highlight our package’s scholarly effort and contributions. First, we not only carefully review the existing packages and (proprietary) code on models for binary, ordered, and unordered dependent variables in our revised paper’s “State of the Field” section but also note that extant software just focuses on estimating standard (non-inflated) discrete choice models. Accordingly, even though researchers across several disciplines often analyze ordered or unordered polytomous dependent variables that exhibit category inflation, there is no full-fledged publicly available package (to our knowledge) that they can use to estimate and assess inflated discrete choice models. Our Python package IDCeMPy, therefore, fills an important lacuna by being the first comprehensive publicly available package that provides an array of functions and a substantial amount of code that fit a wide range of inflated discrete choice models that are applicable across various disciplines. While our IDCeMPy package’s new history results from it being the first full-fledged package that fits a variety of category inflated models for discrete outcomes, it does (as you note) have the potential to be widely used by scholars.

 Second, in the paper’s “Statement of Need” section we list numerous examples of ordered and unordered polytomous dependent variables—that were widely used in the natural and social sciences—that can be analyzed by using the functions in our IDCeMPy package. This suggests (in our view) the wide applicability of the statistical models in IDCeMPy to various disciplines and further highlights the potential empirical contributions our package’s functions and features can provide to researchers with respect to analyzing commonly-used category inflated discrete outcome measures. Third, we emphasize in the revised paper and the package’s documentation that IDCeMPy incorporates features that permit users to assess each included model’s goodness-of-fit, extract the marginal effects of covariates from each model, and conduct Vuong tests for evaluating model performance. We believe that this is a useful contribution since to our knowledge there are no organized functions in extant packages that extract substantive effect from category-inflated discrete choice models and conduct post-estimation diagnostic checks and tests.

Once again, we thank you for your thoughtful feedback.

Regards,
Nguyen
