+++ README.md
@@ -4,8 +4,8 @@
 
 <!-- badges: start -->
 
-[![PyPI version fury.io](https://badge.fury.io/py/idcempy.svg)](https://pypi.org/project/idcempy/0.0.1/)
-[![PyPI pyversions](https://img.shields.io/pypi/pyversions/idcempy.svg)](https://pypi.org/project/idcempy/0.0.1/)
+[![PyPI version fury.io](https://badge.fury.io/py/idcempy.svg)](https://pypi.org/project/idcempy/0.0.5/)
+[![PyPI pyversions](https://img.shields.io/pypi/pyversions/idcempy.svg)](https://pypi.org/project/idcempy/0.0.5/)
 [![Downloads](https://pepy.tech/badge/ziopcpy)](https://pepy.tech/project/ziopcpy)
 [![Project Status: Active â€“ The project has reached a stable, usable state and is being actively developed.](https://www.repostatus.org/badges/latest/active.svg)](https://www.repostatus.org/#active)
 [![License: GPL v3](https://img.shields.io/badge/License-GPLv3-blue.svg)](https://www.gnu.org/licenses/gpl-3.0)
