### 2021-06-29 #3322 @galessiorob
@hknd23

Thank you for your thorough response and instrumentation of @cmaimone's feedback and suggestions. I will review everything in the upcoming days and provide feedback too.

As for the scholarly effort, [JOSS's guidelines](https://joss.readthedocs.io/en/latest/review_criteria.html) point out that while we do ask that the effort, measured as commit history and software age, is no less than three month's of work, these guidelines are not meant to be strictly prescriptive.

I believe this paper checks all the boxes, but I agree that it should be fully functional before being published and well documented.

Thank you for the time and dedication thus far.
