### 2021-06-29 hknd23/idcempy#16 @cmaimone
I'm not seeing the contributing file.  I see a brief Contributions section in the readthedocs documentation and readme.md file, but it doesn't tell people how to submit issues or PRs, report bugs, etc. You'll need to go beyond just telling people to email you for this to really be functional open source software.

Take a look at some of the other published packages on JOSS.  Here's a recent one I grabbed from the website: https://github.com/DrMarc/slab
