### 2021-06-29 hknd23/idcempy#13 @cmaimone
I'm not sure what "in the provided data applications" means.  Do you have code in the repository that runs these timing tests?

I'm also unclear exactly what you're telling us with this information, because there isn't any context.  I'm not sure if this is a long time or short time, because I don't have a comparison point.  I also don't know what would affect the differences in the timing (5 minutes is a lot different than 90!).  A more complex model?  Larger data set?  Number of cores used (is this parallelized)?  What type of computer did you use to run this code?  Could that affect runtime?

My suggestion would be one of the following, but there are other ways you could resolve this.
1. Move the timing information to the examples section of the documentation (out of the paper) to give people a reference for how long it took to run the given example, so they know that if their code has been running for 5 minutes, whether that indicates an error or they should just let it run longer.  You'll still need some additional reference information so they can gauge how the timing on your system might compare to theirs.  Then this becomes a helpful guideline to know people what to expect when running the code instead of a performance claim.
2. If you want to include timing information in the paper, run more systematic timing tests over a variety of models to show how the time to run varies with # observations, # variables, model complexity, type of model, etc.
