### 2021-06-29 hknd23/idcempy#19 @cmaimone
Great!

If you set places=0 though, doesn't this round the values to zero decimal places?  For larger values, this may be appropriate (just comparing the integer part of the number), but for coefficient values < 1, this doesn't seem to be the right choice.  I may be missing something here though.

Move your test files to your test directory.
