### 2021-06-29 #3322 @cmaimone
I reviewed the additional work on this project.  I've updated the relevant issues and the checklist above.  I still have concerns with:

- Performance claims (better, but the timing information isn't clear)
- Contributing guidelines need to be expanded
- Tests (mostly done - great work!, but one issue to check on)

A few other issues remain open that should be resolved, but they are more minor issues.

As a side note, I'd really encourage you to find some additional testers for the package that aren't the reviewers here.  I think that will help you identify any remaining issues with documentation clarity.
