+++ setup.cfg
@@ -1,7 +1,7 @@
 [metadata]
 # replace with your username:
 name = idcempy
-version = 0.0.4
+version = 0.0.5
 author = Nguyen Huynh, Sergio Bejar, Vineeta Yadav, Bumba Mukherjee
 author_email = nguyenhuynh831@gmail.com
 description = Inflated Discrete Choice Models
