### 2021-06-29 6f7d196
added type for data for all functions
