+++ tests/gimnl_unit_test.py
@@ -53,7 +53,7 @@ class TestBimnlLlike(unittest.TestCase):
                          gimnl.bimnl3(pstart_gimnl, x_, x_, yx_, z_,
                                       [0, 2, 1]))
         self.assertAlmostEqual(gimnl.bimnl3(pstart_gimnl, x_, x_, yx_, z_,
-                                            [0, 1, 2]), 2109.38643, places=0)
+                                            [0, 1, 2]), 2109.38643, places=2)
 
 
 class TestSimnlTimnlLlike(unittest.TestCase):
@@ -63,13 +63,13 @@ class TestSimnlTimnlLlike(unittest.TestCase):
                          gimnl.timnl3(pstart_gimnl, x_, x_, yx_, z_,
                                       [2, 1, 0]))
         self.assertAlmostEqual(gimnl.simnl3(pstart_gimnl, x_, x_, yx_, z_,
-                                            [2, 0, 1]), 2242.670607, places=0)
+                                            [2, 0, 1]), 2242.670607, places=2)
         self.assertEqual(gimnl.simnl3(pstart_gimnl, x_, x_, yx_, z_,
                                       [1, 0, 2]),
                          gimnl.timnl3(pstart_gimnl, x_, x_, yx_, z_,
                                       [1, 2, 0]))
         self.assertAlmostEqual(gimnl.timnl3(pstart_gimnl, x_, x_, yx_, z_,
-                                            [1, 2, 0]), 2242.880607, places=0)
+                                            [1, 2, 0]), 2242.880607, places=2)
 
 
 class TestMnlLlike(unittest.TestCase):
@@ -78,7 +78,7 @@ class TestMnlLlike(unittest.TestCase):
                          gimnl.mnl3(pstart_mnl, x_, x_, yx_, [0, 2, 1]))
         self.assertAlmostEqual(gimnl.mnl3(pstart_mnl, x_, x_, yx_, [0, 1, 2]),
                                1401.98008,
-                               places=0)
+                               places=2)
 
 
 class TestMnl(unittest.TestCase):
@@ -88,16 +88,16 @@ class TestMnl(unittest.TestCase):
             40)
         self.assertAlmostEqual(gimnl.mnlmod(
             DAT, x, y, order_Kerry).coefs.iloc[0, 0], -2.189,
-                               places=0)
+                               places=1)
         self.assertAlmostEqual(gimnl.mnlmod(
             DAT, x, y, order_Kerry).coefs.iloc[5, 0], 0.344,
-                               places=0)
+                               places=1)
         self.assertAlmostEqual(gimnl.mnlmod(
             DAT, x, y, order_Bush).coefs.iloc[0, 0], -3.018,
-                               places=0)
+                               places=1)
         self.assertAlmostEqual(gimnl.mnlmod(
             DAT, x, y, order_Bush).coefs.iloc[5, 0], -0.386,
-                               places=0)
+                               places=1)
 
 
 class TestGimnl(unittest.TestCase):
@@ -107,22 +107,22 @@ class TestGimnl(unittest.TestCase):
             45)
         self.assertAlmostEqual(gimnl.gimnlmod(
             DAT, x, y, z, order_Bush, "baseline").coefs.iloc[0, 0], -4.938,
-                               places=0)
+                               places=1)
         self.assertAlmostEqual(gimnl.gimnlmod(
             DAT, x, y, z, order_Bush, "baseline").coefs.iloc[5, 0], -1.632,
-                               places=0)
+                               places=1)
         self.assertAlmostEqual(gimnl.gimnlmod(
             DAT, x, y, z, order_Bush, "baseline").coefs.iloc[10, 0], -0.390,
-                               places=0)
+                               places=1)
[…]
