+++ docs/idcempy_tutorial.rst
@@ -63,13 +63,11 @@ We first import the required libraries, set up the package and import the datase
 .. testcode::
 
   # Import the necessary libraries and package
-
   import pandas as pd
   import urllib
   from idcempy import zmiopc
 
   # Import the "Youth Tobacco Consumption" dataset as a pandas.DataFrame
-
   url='https://github.com/hknd23/zmiopc/blob/main/data/tobacco_cons.csv'
   data = pd.read_csv(url)
 
@@ -216,13 +214,11 @@ We first import the required libraries, set up the package and import the datase
 .. testcode::
 
   # Import the necessary libraries and IDCeMPy.
-
   import pandas as pd
   import urllib
   from idcempy import zmiopc
 
   # Import the "Youth Tobacco Consumption" dataset.
-
   url='https://github.com/hknd23/zmiopc/blob/main/data/tobacco_cons.csv'
 
   # Define a `Pandas` DataFrame.
@@ -357,7 +353,6 @@ We begin by loading the required libraries and IDCeMPy
 .. testcode::
 
   # Import the necessary libraries and IDCeMPy.
-
   import pandas as pd
   import urllib
   from idcempy import zmiopc
@@ -367,11 +362,9 @@ Next, we load the dataset.
 .. testcode::
 
     # Import and read the dataset
-
     url = 'https://github.com/hknd23/idcempy/raw/main/data/EUKnowledge.dta'
 
     # Define a `Pandas` DataFrame
-
     data = pd_read.stata(url)
 
 We then define the lists with the names of the variables used in the model
@@ -501,7 +494,6 @@ First is importing the data and libraries:
 .. testcode::
 
   # Import the necessary libraries and IDCeMPy.
-
   import pandas as pd
   import urllib
   from idcempy import zmiopc
@@ -510,13 +502,11 @@ Next, we load the dataset.
 
 .. testcode::
 
-    # Import and read the dataset
+  # Import and read the dataset
+  url = 'https://github.com/hknd23/idcempy/raw/main/data/EUKnowledge.dta'
 
-    url = 'https://github.com/hknd23/idcempy/raw/main/data/EUKnowledge.dta'
-
-    # Define a `Pandas` DataFrame
-
-    data = pd_read.stata(url)
+  # Define a `Pandas` DataFrame
+  data = pd_read.stata(url)
 
 We then define the lists with the names of the variables used in the model:
 
@@ -575,15 +565,12 @@ The model object :class:`zmiopc.IopCModel` also stores three different diagnosti
 .. testcode::
 
    # Print Log-Likelihood
-
    print(miopc_EU.llik)
 
    # Print AIC
-
    print(miopc_EU.AIC)
 
    # Print VCCOV matrix
-
    rint(miopc_EU.vcov)
 
 To calculate the predicted probabilities:
@@ -591,11 +578,9 @@ To calculate the predicted probabilities:
 .. testcode::
 
    # Define model to fit
-
    fittedmiopc = zmiopc.iopcfit(miopc_EU)
 
    # Print predicted probabilities
-
    print(fittedziopc.responsefull)
 
 The following line of code computes changes in predicted probabilities when the value of a variable changes.
@@ -603,7 +588,6 @@ The following line of code computes changes in predicted probabilities when the 
 .. testcode::
 
    # Define model from which effects will be estimated and number of simulations
-
    miopcxeno = zmiopc.split_effects(miopc_EU, 1, nsims = 10000)
 
 A box plot can illustrate the change in predicted probabilities.
@@ -611,7 +595,6 @@ A box plot can illustrate the change in predicted probabilities.
 .. testcode::
 
     # Get box plot of predicted probabilities
-
     miopcxeno.plot.box(grid='False')
 
 
[…]
