### 2021-06-30 91a0ce2
Update idcempy_tutorial.rst
