+++ docs/idcempy_tutorial.rst
@@ -168,7 +168,6 @@ The following funtion extracts predicted probabilities from the model:
   fittedziop = ziopc.iopfit(ziop_tob)
 
   # Print the predicted probabilities
-
   print(fittedziopc.responsefull)
 
 .. testoutput::
@@ -201,7 +200,6 @@ Results from :func:`zmiopc.split_effects` and :func:`zmiopc.ordered_effects` can
     gender = zmiopc.ordered_effects(ziop_tob, 2, nsims = 10000)
 
     # The box plot from the results:
-
     gender.plot.box(grid='False')
 
 Zero-inflated Ordered Probit (ZiOPC) with Correlated Errors
@@ -285,15 +283,12 @@ To print the estimates of the log-likelihood, AIC, and Variance-Covariance matri
 .. testcode::
 
   # Print Log-Likelihood
-
   print(ziopc_tob.llik)
 
   # Print AIC
-
   print(ziopc_tob.AIC)
 
   # Print VCOV matrix
-
   print(ziopc_tob.vcov)
 
 The AIC of the ziopc_tob model, for example, is:
@@ -307,11 +302,9 @@ The predicted probabilities from the `ziopc_tob` model can be obtained with :fun
 .. testcode::
 
   # Define the model for which you want to estimate the predicted probabilities
-
   fittedziopc = zmiopc.iopcfit(ziopc_tob)
 
   # Print predicted probabilities
-
   print(fittedziopc.responsefull)
 
 .. testoutput::
@@ -329,17 +322,15 @@ The predicted probabilities from the `ziopc_tob` model can be obtained with :fun
 
 .. testcode::
 
-    ziopcgender = zmiopc.split_effects(ziopc_tob, 1, nsims = 10000)
+  ziopcgender = zmiopc.split_effects(ziopc_tob, 1, nsims = 10000)
 
 .. testcode::
 
-   # Calculate change in predicted probabilities
+  # Calculate change in predicted probabilities
+  gender = zmiopc.ordered_effects(ziopc_tob, 1, nsims = 10000)
 
-   gender = zmiopc.ordered_effects(ziopc_tob, 1, nsims = 10000)
-
-   # Box-plot of precicted probabilities
-
-   gender.plot.box(grid='False')
+  # Box-plot of precicted probabilities
+  gender.plot.box(grid='False')
 
 Middle-inflated Ordered Probit (MiOP) without Correlated Errors
 ---------------------------------------------------------------
@@ -361,11 +352,11 @@ Next, we load the dataset.
 
 .. testcode::
 
-    # Import and read the dataset
-    url = 'https://github.com/hknd23/idcempy/raw/main/data/EUKnowledge.dta'
+  # Import and read the dataset
+  url = 'https://github.com/hknd23/idcempy/raw/main/data/EUKnowledge.dta'
 
-    # Define a `Pandas` DataFrame
-    data = pd_read.stata(url)
+  # Define a `Pandas` DataFrame
+  data = pd_read.stata(url)
 
 We then define the lists with the names of the variables used in the model
 
@@ -384,19 +375,18 @@ After importing the dataset and specifying the list of variables from it, the Mi
 
 .. testcode::
 
-   # Model estimation:
+ # Model estimation:
+ miop_EU = zmiopc.iopmod('miop', data, X, Y, Z, method = 'bfgs', weights = 1,offsetx = 0, offsetz = 0)
 
-   miop_EU = zmiopc.iopmod('miop', data, X, Y, Z, method = 'bfgs', weights = 1,offsetx = 0, offsetz = 0)
-
-   # 'miop' = Type of model to be estimated. In this case 'miop'
-   # data = name of Pandas DataFrame
-   # X = variables in the ordered probit stage.
[…]
