### 2021-06-30 fce20fb
Update idcempy_tutorial.rst
