+++ examples/GIMNL_Example.py
@@ -1,7 +1,7 @@
 import pandas as pd
 from idcempy import gimnl
 import os
-import unittest
+import time
 import numpy as np
 
 DAT = pd.read_stata(
