### 2021-07-02 a7f4d10
edits to docs and examples
