+++ examples/GIMNL_Example.py
@@ -17,4 +17,25 @@ y = ['vote_turn']
 order_Kerry = [0, 1, 2]
 order_Bush = [0, 2, 1]
 
-binflatecat = "baseline"
+start_time = time.time()
+modelmnl = gimnl.mnlmod(DAT, x, y, order_Kerry)
+model_time = time.time() - start_time
+print("%s seconds" % model_time)
+
+start_time = time.time()
+model = gimnl.gimnlmod(DAT, x, y, z, order_Kerry, "baseline")
+model_time = time.time() - start_time
+print("%s seconds" % model_time)
+
+x2 = ['educ', 'party7', 'agegroup2']
+z2 = ['educ', 'agegroup2']
+
+start_time = time.time()
+model_small = gimnl.gimnlmod(DAT, x2, y, z2, order_Kerry, "baseline")
+model_time = time.time() - start_time
+print("%s seconds" % model_time)
+
+start_time = time.time()
+mnl_small = gimnl.mnlmod(DAT, x2, y, order_Kerry)
+model_time = time.time() - start_time
+print("%s seconds" % model_time)
