+++ CONTRIBUTING.md
@@ -1,0 +1,145 @@
+# Contributing to ICDeMPy
+
+Welcome to `ICDeMPy` package, and thank you for considering contributing to this Python
+package. Please note that you can contribute in many ways, including fixing a bug, adding a new
+feature, code patches, editing the documentation, and patch reviews. We appreciate your help in
+improving any feature included in this package.
+
+This contributing guide contains the necessary information you need to know to help build and
+improve `ICDeMPy`. Please read and follow these guidelines as it will make both communication
+and the contribution process easy and effective. We will reciprocate by [bug reports](#fixing bugs), [feature requests](#evaluating changes), and helping you finalize your [pull requests](#pull requests).
+
+## Code of Conduct
+
+We treat the open-source community with respect and hold ourselves as well as other
+contributors to high standards of communication. By contributing to this project, you agree to
+uphold our following [Code of Conduct](https://github.com/auth0/open-source-template/blob/master/CODE-OF-CONDUCT.md)
+
+- Focusing on what is best for the community
+- Respect differing viewpoints and accept constructive criticisms
+- Avoid conduct that could reasonably be considered inappropriate in a professional setting
+
+Our responsibilities as maintainers include the right to remove, edit, or reject comments,
+commits, code, issues, and other contributions that are not aligned to this **Code of Conduct**.
+
+## Bug Reports
+
+We use GitHub issues to track public bugs. Note that a bug can be reported by simply opening a
+new issue. A useful Bug Report is one that includes:
+
+- A summary or background.
+- Steps to reproduce. Please be as specific as possible in this regard.
+- Give sample code if you can. This makes it easier to understand, track and correct the
+main issues that you are raising.
+- Briefly explain what you expected would happen and what actually transpired (contrary
+to your expectations).
+- Finally, please note the steps you may have undertaken to address the bug.
+
+Below you will find an example of the anatomy of a useful bug report. 
+
+Example:
+
+> Title: Short and descriptive example bug report title
+>
+> Description: A summary of the issue and the browser/OS environment in which it occurs. If
+> suitable, include the steps required to reproduce the bug as follows.
+>
+> 1. This is the first step
+> 2. This is the second step
+> 3. Further steps, etc.
+>
+> `<url>` - a link to the reduced test case
+>
+> Additional Information: Any other information you want to share that is relevant to the issue being
+> reported. This might include the lines of code that you have identified as
+> causing the bug, and potential solutions (and your opinions on their
+> merits).
+
+> Name:  Your name and preferred method of contact. 
+
+## Feature Requests
+
+We welcome any Feature Requests. That said, please ensure that your request to add a new
+feature, for example, fits the objectives and scope of the `IDCeMPy`. Please provide as much
+detail as possible and ensure that the feature that you intend to add is compatible with other
+features included in this package.
+
+## Issues
+
+Contributors should use Issues to report problems with the library, request a new feature, or to
+discuss potential changes before a PR is created. When a contributor creates a new Issue, a
+template will be loaded that will guide him or her to collect and provide the necessary
+information that we (the maintainers) need to investigate. If the contributor finds an Issue that
+addresses the problem they are having, then he or she should add their own reproduction
+information to the existing issue rather than creating a new one.
+
+## Pull Requests
+
+Pull requests to our libraries in `IDCeMPy` are more than welcome. Pull requests (PRs) are the
+best way to propose changes to the codebase. In general, we use [GitHub flow](https://guides.github.com/introduction/flow/index.html) for PRs to:
+
[…]
