### 2021-07-03 66ce749
Delete CONTRIBUTING.md
