+++ docs/CONTRIBUTING.md
@@ -1,8 +1,0 @@
-# How to request support
-
-For questions about IDCeMPy, please contact <nkh8@psu.edu> or open a new issue. 
-
-# How to contribute
-
-We welcome contributors to IDCeMPy. Issues can be raised by any user to report bugs. 
-For enhancement suggestions such as improvements to functionality or new features, please email <nkh8@psu.edu> for prompt response.
