### 2021-07-03 a95d20e
change version for up-to-date readme (package uploaded to PyPI)
