+++ paper.md
@@ -43,7 +43,9 @@ The software written in R or LIMDEP and the STATA commands listed above are undo
 
 # Statement of Need 
 
-Although our `IDCeMPy` package also fits standard discrete choice models, what makes our package unique is that unlike existing software, it offers functions to fit and assess the performance of both Zero-Inflated and Middle-Inflated Ordered Probit (OP) models without and with correlated errors as well as a set of Generalized-Inflated MNL models. The models included in `IDCeMPy` account for the excessive proportion of observations in any given ordered or unordered outcome category by combining a single binary probit or logit split-stage equation with either an ordered probit outcome stage (for the Zero and Middle-Inflated OP models) or a MNL outcome-stage equation. Users can treat the error terms from the two equations in the Zero and Middle-Inflated OP models as independent or correlated in the package’s estimation routines. `IDCeMPy` also provides functions to assess each included model’s goodness-of-fit via the AIC statistics, extract the covariates’ marginal effects from each model, and conduct Vuong tests for comparing the performance between the standard and inflated discrete choice models. The functions in `IDCeMPy` use quasi-Newton optimization methods such as the Broyden-Fletcher-Goldfarb-Shanno algorithm for Maximum-Likelihood-Estimation (MLE), which facilitates convergence and estimation speed. For instance, the run-time for estimating the inflated discrete choice models without correlated errors in the provided data applications using `IDCeMPy` range from approximately 8 to 75 seconds, while the correlated error models run-time vary from about 5 to almost 90 minutes. Another feature is that the coefficients, standard errors, and confidence intervals obtained for each model estimated in `IDCeMPy` are in `pandas.DataFrame` [@mckinney-proc-scipy-2010] format and are stored as class attribute `.coefs`. This allows for easy export to csv or excel, which makes it easier for users to perform diagnostic tests and extract marginal effects. `IDCeMPy` is thus important as it provides a much-needed unified software package to fit statistical models to account for category inflation in several ordered and unordered outcome variables used across fields as diverse as Economics, Engineering, Marketing, Political Science, Public Health, Sociology, and Transportation research. Users can employ the wide range of statistical models in `IDCeMPy` to assess, 
+Although our `IDCeMPy` package also fits standard discrete choice models, what makes our package unique is that unlike existing software, it offers functions to fit and assess the performance of both Zero-Inflated and Middle-Inflated Ordered Probit (OP) models without and with correlated errors as well as a set of Generalized-Inflated MNL models. The models included in `IDCeMPy` account for the excessive proportion of observations in any given ordered or unordered outcome category by combining a single binary probit or logit split-stage equation with either an ordered probit outcome stage (for the Zero and Middle-Inflated OP models) or a MNL outcome-stage equation. Users can treat the error terms from the two equations in the Zero and Middle-Inflated OP models as independent or correlated in the package’s estimation routines. `IDCeMPy` also provides functions to assess each included model’s goodness-of-fit via the AIC statistics, extract the covariates’ marginal effects from each model, and conduct Vuong tests for comparing the performance between the standard and inflated discrete choice models. 
+
[…]
