### 2021-07-04 b7a930f
Update bug_report.md
