+++ .github/ISSUE_TEMPLATE/bug_report.md
@@ -6,6 +6,8 @@ labels: bug
 assignees: ''
 
 ---
+**Expected behavior**
+A clear and concise description of what you expected to happen.
 
 **Describe the bug**
 A clear and concise description of what the bug is.
@@ -15,9 +17,6 @@ Steps to reproduce the behavior:
 1. Import '...'
 2. Run function '....'
 3. See error
-
-**Expected behavior**
-A clear and concise description of what you expected to happen.
 
 **Screenshots**
 If applicable, add screenshots to help explain your problem.
