### 2021-07-04 d6ee4d5
Update bug_report.md
