+++ .github/ISSUE_TEMPLATE/bug_report.md
@@ -6,17 +6,22 @@ labels: bug
 assignees: ''
 
 ---
-**Expected behavior**
-A clear and concise description of what you expected to happen.
-
 **Describe the bug**
 A clear and concise description of what the bug is.
+
+**Expected behavior**
+A clear and concise description of what you expected to happen. 
+
+**Actual behavior**
+A clear and concise description of the output that you received. 
 
 **To Reproduce**
 Steps to reproduce the behavior:
 1. Import '...'
 2. Run function '....'
 3. See error
+
+Ideally, the steps to reproduce the bugs should reproduce exactly the actual behavior. If applicable, include URL to data source. 
 
 **Screenshots**
 If applicable, add screenshots to help explain your problem.
