+++ CONTRIBUTING.md
@@ -25,11 +25,13 @@ commits, code, issues, and other contributions that are not aligned to this **Co
 ## Bug Reports
 
 We use GitHub issues to track public bugs. Note that a bug can be reported by simply opening a
-new issue. An issue template for reporting bugs is available. A useful Bug Report is one that includes:
+new issue. An issue template for reporting bugs is available. Before openning a new issue,
+please check whether a similar issue has already been reported.
+A useful Bug Report is one that includes:
 
 - A summary or background.
 - Steps to reproduce. Please be as specific as possible in this regard.
-- Give sample code if you can. This makes it easier to understand, track and correct the
+- Give sample code and screenshots if you can. This makes it easier to understand, track and correct the
 main issues that you are raising.
 - Briefly explain what you expected would happen and what actually transpired (contrary
 to your expectations).
@@ -37,11 +39,11 @@ to your expectations).
 
 ## Feature Requests
 
-We welcome any requests to add new features. That said, please ensure that your request to add a new
+We welcome any requests to add new features. The feature request issue template is
+available to facilitate the process. That said, please ensure that your request to add a new
 feature, for example, fits the objectives and scope of the `IDCeMPy`.
-Please use the feature request template when creating an issue and provide as much
-detail as possible and ensure that the feature that you intend to add is compatible with other
-features included in this package.
+Please provide as much detail as possible and ensure that the feature 
+that you intend to add is compatible with other features included in this package. 
 
 ## Issues
 
