### 2021-07-05 fa16078
Update idcempy_tutorial.rst
