+++ paper.bib
@@ -253,16 +253,17 @@
   @article{henningsen2014micecon,
   title={micEcon: Microeconomic Analysis and Modelling},
   author={Henningsen, A},
-  journal={R package version 0.6-12, URL http://CRAN. R-project. org/package= micEcon},
+  url = {https://cran.r-project.org/package=micEcon},
+  journal={R package version 0.6-12},
   year={2014}
 }
 
 @article{croissant2012estimation,
   title={Estimation of multinomial logit models in R: The mlogit Packages},
   author={Croissant, Yves},
-  journal={R package version 0.2-2. URL: http://cran. r-project. org/web/packages/mlogit/vignettes/mlogit. pdf},
+  url = {https://cran.r-project.org/package=mlogitl},
+  journal={R package version 0.2-2.},
   year={2012},
-  publisher={Citeseer}
 }
 
 @article{hasan2014fast,
@@ -279,8 +280,8 @@
 @article{sarrias2015gmnl,
   title={gmnl: Multinomial Logit Models with Random Parameters},
   author={Sarrias, M and Daziano, R},
-  journal={R package version},
-  volume={1},
+  url = {https://cran.r-project.org/package=gmnl},
+  journal={R package version 1.1-3.2},
   year={2015}
 }
 
