+++ docs/idcempy_tutorial.rst
@@ -33,7 +33,7 @@ An excessive (“inflated”) share of observations—stemming from two distinct
 Each inflated discrete choice model in this package addresses category inflation in one’s discrete outcome—unordered or unordered polytomous—of interest by jointly estimating a binary split-stage equation and an ordered or multinomial discrete choice outcome equation.
 
 Installation
-=============
+============
 The package can be installed in two different ways:
 
 1. From `PyPi <https://pypi.org/project/idcempy/>`__:
@@ -60,7 +60,7 @@ Please note that for models in the `zmiopc` module, the run-time for models with
 than their without correlated errors counterparts using :func:`zmiopc.iopmod`. Other factors affecting run-time are the number of observations and the number of covariates.
 
 Zero-inflated Ordered Probit (ZiOP) Model without Correlated Errors
---------------------------------------------------------------------
+-------------------------------------------------------------------
 The :func:`zmiopc.iopmod` function estimates regression objects for "zero-inflated" and "middle-inflated" ordered probit models without correlated errors.
 This section provides instruction to estimate the ZiOP model using the self-reported smoking behavior as empirical example.
 
@@ -584,7 +584,8 @@ A box plot can illustrate the change in predicted probabilities.
     miopcxeno.plot.box(grid='False')
 
 
-To calculate the change in predicted probabilities of the outcome variable in the outcome-stage when the value of a covarariate changes. The box plots below display the change in predicted probabilities of the outcome variable in the MiOPC model estimated above when Xenophobia increases one standard deviation from its mean value.
+To calculate the change in predicted probabilities of the outcome variable in the outcome-stage when the value of a covarariate changes.
+The box plots below display the change in predicted probabilities of the outcome variable in the MiOPC model estimated above when Xenophobia increases one standard deviation from its mean value.
 
 .. testcode::
 
@@ -688,7 +689,8 @@ Log-likelihood, AIC, and Variance-Covariance matrix can be extracted with:
 The Vuong Test
 --------------
 
-Harris and Zhao (`2007 <https://doi.org/10.1016/j.jeconom.2007.01.002>`__) suggest that a variant of the Vuong (`1989 <https://www.jstor.org/stable/1912557>`__) Test (with a v statistic) can be used to compare the performance of the ZiOP versus the standard Ordered Probit (OP) model. The Vuong's test formula is:
+Harris and Zhao (`2007 <https://doi.org/10.1016/j.jeconom.2007.01.002>`__) suggest that a variant of the Vuong (`1989 <https://www.jstor.org/stable/1912557>`__)
+Test (with a v statistic) can be used to compare the performance of the ZiOP versus the standard Ordered Probit (OP) model. The Vuong's test formula is:
 
 .. math::
 
@@ -696,7 +698,8 @@ Harris and Zhao (`2007 <https://doi.org/10.1016/j.jeconom.2007.01.002>`__) sugge
 
 where v < -1.96 favors the more general (ZiOP/ZiOPC) model, -1.96 < v < 1.96 lends no support to either model, and v > 1.96 supports the simpler (OP) model.
 
-The OP and ZiOP models must have the same number of observations, and the OP must have the same number of covariates as ZiOP's OP stage. The statistic below reveals that the OP model is preferred over the ZiOP model.
+The OP and ZiOP models must have the same number of observations, and the OP must have the same number of covariates as ZiOP's OP stage.
+The statistic below reveals that the OP model is preferred over the ZiOP model.
 
 .. testcode::
 
@@ -730,7 +733,8 @@ Similar to the models in the :py:mod:`zmiopc` module, the first step is to impor
   # Define a `Pandas` DataFrame
   data = pd.read_stata(url)
 
[…]
