+++ paper.bib
@@ -196,7 +196,7 @@
 }
 
 @article{sarrias2016discrete,
-  title={Discrete choice models with random parameters in R: The Rchoice Package},
+  title={Discrete choice models with random parameters in {R}: The Rchoice Package},
   author={Sarrias, Mauricio},
   journal={Journal of Statistical Software},
   volume={74},
@@ -259,7 +259,7 @@
 }
 
 @article{croissant2012estimation,
-  title={Estimation of multinomial logit models in R: The mlogit Packages},
+  title={Estimation of multinomial logit models in {R}: The mlogit Packages},
   author={Croissant, Yves},
   url = {https://cran.r-project.org/package=mlogitl},
   journal={R package version 0.2-2.},
