+++ docs/idcempy_tutorial.rst
@@ -112,6 +112,7 @@ Results from the model:
 The following message will appear when the model has converged:
 
 .. testoutput::
+
          Warning: Desired error not necessarily achieved due to precision loss.
          Current function value: 5060.160903
          Iterations: 79
@@ -124,9 +125,11 @@ Object :class:`zmiopc.IopModel` stores model results and goodness-of-fit tests i
 The following line of code prints the estimates of coefficients:
 
 .. testcode::
+
    print(ziop_tob.coefs)
 
 .. testoutput::
+
                             Coef        SE      tscore        p           2.5%      97.5%
    cut1                   1.693797  0.054383  31.145912  0.000000e+00   1.587207   1.800387
    cut2                  -0.757830  0.032290 -23.469359  0.000000e+00  -0.821119  -0.694542
@@ -144,6 +147,7 @@ The model object :class:`zmiopc.IopModel` also stores three different diagnostic
 They can be obtained via the following:
 
 .. testcode::
+
   print(ziop_tob.llik)
   print(ziop_tob.AIC)
   print(ziop_tob.vcov)
@@ -151,21 +155,25 @@ They can be obtained via the following:
 An example for the AIC:
 
 .. testcode::
+
   print(ziop_tob.AIC)
 
 .. testoutput::
+
   10138.321806674261
 
 The following funtion extracts predicted probabilities from the model:
 :func:`zmiopc.iopfit` returns :class:`zmiopc.FittedVals` containing fitted probablities.
 
 .. testcode::
+
   fittedziop = ziopc.iopfit(ziop_tob)
 
   # Print the predicted probabilities
   print(fittedziopc.responsefull)
 
 .. testoutput::
+
   array[[0.8822262  0.06879832 0.01455244 0.0242539  0.01016914]
  [0.84619828 0.08041296 0.01916279 0.03549797 0.01872801]
  [0.93105632 0.04349743 0.00831396 0.0127043  0.004428  ]
