### 2021-07-05 68ebca6
Update idcempy_tutorial.rst
