+++ paper.bib
@@ -314,5 +314,14 @@
   doi       = { 10.25080/Majora-92bf1922-00a }
 }  
   
-  
-  
+  @article{sarrias2017multinomial,
+  title={Multinomial logit models with continuous and discrete individual heterogeneity in {R}: the gmnl package},
+  author={Sarrias, Mauricio and Daziano, Ricardo and others},
+  journal={Journal of Statistical Software},
+  volume={79},
+  number={2},
+  pages={1--46},
+  year={2017},
+ doi = {10.18637/jss.v079.i02}
+  publisher={Foundation for Open Access Statistics}
+}
