+++ paper.bib
@@ -194,23 +194,11 @@
   publisher={Cambridge University Press},
   doi = {10.1017/CBO9780511845062}
 }
-
-@article{sarrias2016discrete,
-  title={Discrete choice models with random parameters in {R}: The Rchoice Package},
-  author={Sarrias, Mauricio},
-  journal={Journal of Statistical Software},
-  volume={74},
-  number={10},
-  pages={1--31},
-  year={2016},
-  publisher={Foundation for Open Access Statistics},
-  doi = {10.18637/jss.v074.i10}
-  }
   
   @article{JSSv074i10,
-   title = {Discrete Choice Models with Random Parameters in R: The Rchoice Package},
+   title = {Discrete Choice Models with Random Parameters in {R}: The {R}choice Package},
    author = {Sarrias, Mauricio},
-   journal = {Journal of Statistical Software, Articles},
+   journal = {Journal of Statistical Software},
    volume = {74},
    number = {10},
    year = {2016},
@@ -267,7 +255,7 @@
 }
 
 @article{hasan2014fast,
-  title={Fast estimation of multinomial logit models: R package mnlogit},
+  title={Fast estimation of multinomial logit models: {R} package mnlogit},
   author={Hasan, Asad and Zhiyu, Wang and Mahani, Alireza S},
   journal={Journal of Statistical Software},
   volume={75},
@@ -286,7 +274,7 @@
 }
 
 @article{liverani2015premium,
-  title={PReMiuM: An R package for profile regression mixture models using Dirichlet processes},
+  title={{PR}e{M}iu{M}: An {R} package for profile regression mixture models using Dirichlet processes},
   author={Liverani, Silvia and Hastie, David I and Azizi, Lamiae and Papathomas, Michail and Richardson, Sylvia},
   journal={Journal of statistical software},
   volume={64},
@@ -298,7 +286,7 @@
   }
   
   @article{bierlaire2016pythonbiogeme,
-  title={PythonBiogeme: a short introduction. Report TRANSP-OR 160706, Series on Biogeme},
+  title={{P}ython{B}iogeme: a short introduction. Report TRANSP-OR 160706, Series on Biogeme},
   author={Bierlaire, M},
   journal={Transport and Mobility Laboratory, School of Architecture, Civil and Environmental Engineering, Ecole Polytechnique F{\'e}d{\'e}rale de Lausanne, Switzerland},
   year={2016}
