+++ paper.bib
@@ -310,6 +310,6 @@
   number={2},
   pages={1--46},
   year={2017},
- doi = {10.18637/jss.v079.i02}
+ doi = {10.18637/jss.v079.i02},
   publisher={Foundation for Open Access Statistics}
 }
