+++ paper.bib
@@ -247,7 +247,7 @@
 }
 
 @article{croissant2012estimation,
-  title={Estimation of multinomial logit models in {R}: The mlogit Packages},
+  title={mlogit: Multinomial Logit Models},
   author={Croissant, Yves},
   url = {https://cran.r-project.org/package=mlogitl},
   journal={R package version 0.2-2.},
