### 2021-07-06 bb48b43
Update idcempy_tutorial.rst
