### 2021-07-06 hknd23/idcempy#13 @hknd23
We addressed these crucial points as follows. First, as per your suggestion, we have added code from the time module for the examples in our documentation that provides the run-time in seconds for every inflated discrete choice example model that is estimated using the relevant function in our IDCeMPy package. Specifically, the code mentioned above, which is available in the [/examples](https://github.com/hknd23/idcempy/tree/main/examples) directory, tracks and reports the run-time—which users can observe—for each estimated Zero-Inflated as well as Middle-Inflated ordered probit model with or without correlated errors [ZiOP(C) and MiOP(C)], and the General-Inflated MNL [i.e., GiMNL] model. Second, as described below in our next response, we report in our documentation the exact time taken to fit each ZiOP, ZiOPC, MiOP, MiOPC, and GiMNL specification (we estimated several of these specifications) using the relevant code in our package. This provides a reference point for users to gauge the time taken to estimate each model employing the example datasets that we provide. Third, we use INTEL i-7 Quad-Core (16 GB RAM) to estimate the models in the package, which is explicitly stated in the examples included in our documentation. We also note that the models are fit using one core at a time and are therefore not parallelized. Information about the standard computer specs that we employ for our analysis may serve as a helpful guideline for users when running the code included in IDCeMPy. We next turn to address the remaining queries you raised about the runtime for the models in our package.

These are great suggestions that have been incorporated in the following manner. We first followed your advice by removing information—that is, deleting sentences—from our paper in which we had broadly suggested performance claims by mentioning the estimation time for our inflated discrete choice models. Instead, in the examples section of our revised documentation (but not the paper), we report the results from several distinct ZiOP, ZiOPC, MiOP, MiOPC, and GiMNL specifications as well as the run-time for each of these models. For instance, we report the run-time for the Zero-inflated Ordered Probit model with and without correlated errors (ZiOP/ZiOPC models) that have been estimated on two separate datasets: the repression dataset which contains 1984 observations, and the larger youth national tobacco consumption dataset that has 9624 observations. In the case of the Middle-inflated Ordered Probit model with and without correlated errors, we present the results and run-time for several different MiOP and MiOPC specifications: baseline specifications that include just four covariates, "intermediate" specifications that include ten covariates, and a full-fledged or complete specification that incorporates twenty-seven covariates. We also present the estimates and run-time for a full-fledged GiMNL model that includes twenty-three covariates. As we mentioned earlier, the code and data are available on the repository ([/examples](https://github.com/hknd23/idcempy/tree/main/examples) and [/data](https://github.com/hknd23/idcempy/tree/main/data)).
