### 2021-07-06 hknd23/idcempy#16 @hknd23
We apologize for not thoroughly addressing your point about the contributing file (in other words, contribution guide) in our previous round of revisions. Thank you for providing the link to the contributing file in a publicly available Github repo. This is an extremely helpful example. We have added a detailed contribution guide to our repo using both the template from your suggested link and other available templates for contributing files as a reference point. Our revised and updated contribution guide is included in the readthedocs documentation and readme.md file available at https://github.com/hknd23/idcempy/blob/main/CONTRIBUTING.md. This link provides details about the contents included in the contribution guide associated with our IDCeMPy package. Stated briefly, however, the updated contribution guide is divided into the following five categories: Code of Conduct, Bug Reports, Features Requests, Pull Requests (PRs), License, Getting Help. Detailed information and guidance to potential contributors on how to report bugs, provide PRs as well as Features requests and so on is provided in the relevant categories mentioned above. In addition, we have added issue templates for bug reports and feature requests to facilitate the processes.
