### 2021-07-06 hknd23/idcempy#19 @hknd23
This is an important question. When we set places = 0, the tests will pass for coefficients with differences of up to 1 with their respective asserted values. It is thus plausible that setting places = 0 may not be appropriate for smaller coefficients obtained (for example) from the data we use for our tests. However, to ensure that the functions in our package provide results that converge approximately to the asserted values, we went over all our tests and made the adjustments for them to be more restrictive. For instance, we set places = 1 for tests to compare coefficients. For tests for the likelihood and Vuong test functions, we set places = 2. Doing so will make the tests more sensitive and likely when the functions’ results are too different from the asserted values. Following your request, the test files have been moved to our [/test](https://github.com/hknd23/idcempy/tree/main/tests) directory.
