### 2021-07-06 #3322 @galessiorob
Based on @cmaimone's suggestion I asked @tmickleydoyle to test the package first as an external user and have now asked him to do a formal review. (Thanks for volunteering @tmickleydoyle 🎉 )

You can find his notes on installation and usage [here.](https://gist.github.com/tmickleydoyle/dbe586a6288879124ad3f551028cb6f6)
