### 2021-07-06 #3322 @galessiorob
@tmickleydoyle please find [your personal checklist](https://github.com/openjournals/joss-reviews/issues/3322#issue-907576536) at the top. Feel free to open individual issues for @hknd23 to work through or mention them in this thread. Thanks again for reviewing this paper, let me know if you have any questions about the process 😄
