### 2021-07-06 993f02d
Update idcempy_tutorial.rst
