### 2021-07-06 17e86fd
Update idcempy_tutorial.rst
