### 2021-07-06 #3322 @cmaimone
I've closed all of the remaining issues I had opened on the repo.  @hknd23 I think all of your work will make the package a lot easier for people to use (and feel confident in), which will hopefully also lead to more citations and use.  I'm good with the publication moving forward after @tmickleydoyle finishes testing/review.
