### 2021-07-11 #3322 @tmickleydoyle
Hey, @cmaimone 👋  Thank you all for including me in the review of this potentially very useful package 🎉  I have completed my checklist in the initial comment. Best of luck 🥳
