### 2021-07-13 #3322 @galessiorob
@tmickleydoyle @cmaimone and @jungtaekkim Thank you so much for your time and thorough reviews! This paper and package has come a long way and is almost ready to publish 🚀

@hknd23 I have some final small changes and suggestions:
- [Readme.md edits](https://github.com/hknd23/idcempy/pull/22)
- [Paper edits](https://github.com/hknd23/idcempy/pull/23)
