+++ setup.cfg
@@ -1,7 +1,7 @@
 [metadata]
 # replace with your username:
 name = idcempy
-version = 0.0.9
+version = 0.1.0
 author = Nguyen Huynh, Sergio Bejar, Vineeta Yadav, Bumba Mukherjee
 author_email = nguyenhuynh831@gmail.com
 description = Inflated Discrete Choice Models
@@ -11,8 +11,10 @@ url = https://github.com/hknd23/idcempy
 classifiers =
 	Development Status :: 3 - Alpha
 	License :: OSI Approved :: GNU General Public License v3 (GPLv3)
-        Programming Language :: Python :: 3.7
-        Intended Audience :: Science/Research
+    Programming Language :: Python :: 3.7
+    Programming Language :: Python :: 3.8
+    Programming Language :: Python :: 3.9
+    Intended Audience :: Science/Research
 [options]
 packages = find:
 python_requires = >=3.7
