### 2021-07-15 #3322 @hknd23
Hi @galessiorob, we have given the paper one final round of proofreading and made some small edits. We have updated the package's latest version to 0.1.1 as well. Thank you very much for your great help throughout the review process. Please let us know if there is anything else you need.
