### 2021-07-18 #3322 @galessiorob
@hknd23 last thing we need from you is to deposit your software in Zenodo and post the DOI here. After that, we can set the final version, generate a final proof, and have an Editor in Chief do the publishing. 🚀
