### 2021-07-20 #3322 @danielskatz
@hknd23 - I've suggested a bunch of changes in https://github.com/hknd23/idcempy/pull/24.  Also, I note entries in the bib that aren't in the right case - you can use extra {}s to protect cases in these entries.

e.g. The routledge handbook of agricultural economics should be The Routledge Handbook of Agricultural Economics; Modern applied statistics with s should be Modern Applied Statistics with S (though I'm not sure about the S at the end as I don't know this journal).
