### 2021-07-20 #3322 @danielskatz
@hknd23 - I see you've merged the changes - once you've made the bib changes, please use `@whedon generate pdf` to regenerate the paper, and let me know when you are done, then we can proceed with acceptance
