+++ paper.bib
@@ -33,7 +33,7 @@
 @article{crossley2002reliability,
   title={The reliability of self-assessed health status},
   author={Crossley, Thomas F and Kennedy, Steven},
-  journal={Journal of health economics},
+  journal={Journal of Health Economics},
   volume={21},
   number={4},
   pages={643--658},
@@ -45,7 +45,7 @@
 @article{baker2004self,
   title={What do self-reported, objective, measures of health measure?},
   author={Baker, Michael and Stabile, Mark and Deri, Catherine},
-  journal={Journal of human Resources},
+  journal={Journal of Human Resources},
   volume={39},
   number={4},
   pages={1067--1093},
@@ -148,7 +148,7 @@
 @article{fountas2018analysis,
   title={Analysis of accident injury-severities using a correlated random parameters ordered probit approach with time variant covariates},
   author={Fountas, Grigorios and Anastasopoulos, Panagiotis Ch and Abdel-Aty, Mohamed},
-  journal={Analytic methods in accident research},
+  journal={Analytic Methods in Accident Research},
   volume={18},
   pages={57--68},
   year={2018},
@@ -231,7 +231,7 @@
 @incollection{venables2002random,
   title={Random and mixed effects},
   author={Venables, WN and Ripley, BD},
-  booktitle={Modern applied statistics with S},
+  booktitle={Modern Applied Statistics with S},
   pages={271--300},
   year={2002},
   publisher={Springer},
@@ -242,7 +242,7 @@
   title={micEcon: Microeconomic Analysis and Modelling},
   author={Henningsen, A},
   url = {https://cran.r-project.org/package=micEcon},
-  journal={R package version 0.6-12},
+  journal={R Package Version 0.6-12},
   year={2014}
 }
 
@@ -250,7 +250,7 @@
   title={mlogit: Multinomial Logit Models},
   author={Croissant, Yves},
   url = {https://cran.r-project.org/package=mlogitl},
-  journal={R package version 0.2-2.},
+  journal={R Package Version 0.2-2.},
   year={2012},
 }
 
@@ -269,14 +269,14 @@
   title={gmnl: Multinomial Logit Models with Random Parameters},
   author={Sarrias, M and Daziano, R},
   url = {https://cran.r-project.org/package=gmnl},
-  journal={R package version 1.1-3.2},
+  journal={R Package Version 1.1-3.2},
   year={2015}
 }
 
 @article{liverani2015premium,
   title={{PR}e{M}iu{M}: An {R} package for profile regression mixture models using Dirichlet processes},
   author={Liverani, Silvia and Hastie, David I and Azizi, Lamiae and Papathomas, Michail and Richardson, Sylvia},
-  journal={Journal of statistical software},
+  journal={Journal of Statistical Software},
   volume={64},
   number={7},
   pages={1},
@@ -285,13 +285,12 @@
   doi = {10.18637/jss.v064.i07}  
   }
   
-  @article{bierlaire2016pythonbiogeme,
-  title={{P}ython{B}iogeme: a short introduction. Report TRANSP-OR 160706, Series on Biogeme},
-  author={Bierlaire, M},
-  journal={Transport and Mobility Laboratory, School of Architecture, Civil and Environmental Engineering, Ecole Polytechnique F{\'e}d{\'e}rale de Lausanne, Switzerland},
+  @techreport{bierlaire2016pythonbiogeme,
+  title={{P}ython{B}iogeme: a short introduction},
+  author={Bierlaire, Michel},
   year={2016}
 }  
-  
+
   @InProceedings{ mckinney-proc-scipy-2010,
   author    = { {W}es {M}c{K}inney },
   title     = { {D}ata {S}tructures for {S}tatistical {C}omputing in {P}ython },
