+++ paper.bib
@@ -41,7 +41,6 @@
   publisher={Elsevier}
 }
 
-
 @article{baker2004self,
   title={What do self-reported, objective, measures of health measure?},
   author={Baker, Michael and Stabile, Mark and Deri, Catherine},
@@ -53,8 +52,6 @@
   publisher={University of Wisconsin Press},
   doi = {10.3386/w8419}
 }
-
-
 
 @article{senese1997between,
   title={Between dispute and war: The effect of joint democracy on interstate conflict escalation},
@@ -156,13 +153,6 @@
   doi = {10.1016/j.amar.2018.04.003}
 }
 
-@techreport{richards2018new,
-  title={New empirical models in consumer demand},
-  author={Richards, Timothy and Bonnet, Celine and others},
-  year={2018},
-  doi = {10.4324/9781315623351-27}
-}
-
 @article{elgun2007exposure,
   title={Exposure to European Union policies and support for membership in the candidate countries},
   author={Elg{\"u}n, {\"O}zlem and Tillman, Erik R},
@@ -209,7 +199,7 @@
   @incollection{richards2018new,
   title={New empirical models in consumer demand},
   author={Richards, Timothy J and Bonnet, Celine},
-  booktitle={The Routledge Handbook of Agricultural Economics},
+  booktitle={{T}he {R}outledge {H}andbook of {A}gricultural {E}conomics},
   pages={488--511},
   year={2018},
   publisher={Routledge},
