+++ paper.bib
@@ -221,7 +221,7 @@
 @incollection{venables2002random,
   title={Random and mixed effects},
   author={Venables, WN and Ripley, BD},
-  booktitle={Modern Applied Statistics with S},
+  booktitle={{M}odern {A}pplied {S}tatistics with {S}},
   pages={271--300},
   year={2002},
   publisher={Springer},
