+++ paper.bib
@@ -118,18 +118,6 @@
   doi = {10.1111/j.1540-5907.2009.00399.x}
 }
 
-@article{brown2020modelling,
-  title={Modelling Category Inflation with Multiple Inflation Processes: Estimation, Specification and Testing 1},
-  author={Brown, Sarah and Harris, Mark N and Spencer, Christopher},
-  journal={Oxford Bulletin of Economics and Statistics},
-  volume={82},
-  number={6},
-  pages={1342--1361},
-  year={2020},
-  publisher={Wiley Online Library},
-  doi = {10.1111/obes.12366}
-}
-
 @article{xia2019gidm,
   title={gidm: A command for generalized inflated discrete models},
   author={Xia, Yiwei and Zhou, Yisu and Cai, Tianji},
