### 2021-07-20 #3322 @hknd23
@danielskatz Thank you very much for your suggestions. We have reviewd and  incorporated them to the paper. We also capitalized the journal and book names in the .bib file and removed the duplicated entries. Please let us know if there is anything else you need.
